% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{absolute}
\alias{absolute}
\title{Check if File Path is Absolute}
\usage{
absolute(x)
}
\arguments{
\item{x}{character (a file path)}
}
\value{
logical; TRUE if x starts with / or .: (e.g. C:)
}
\description{
Checks if file path is absolute.
}
