% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{as.best.default}
\alias{as.best.default}
\title{Convert Vector to Best of Numeric or Character}
\usage{
\method{as.best}{default}(x, prefix = "", na.strings = c(".", "NA", ""), ...)
}
\arguments{
\item{x}{default}

\item{prefix}{prefix for viewing numerics}

\item{na.strings}{strings that should be treated as NA}

\item{...}{ignored}
}
\description{
Convert vector to best of numeric or character.
}
\details{
Converts vector to numeric if doing so creates no new NA; otherwise to character. Number-like vectors that are nevertheless character are prefixed by default to make this explicit when viewing only a few rows.
}
\seealso{
Other util: 
\code{\link{\%contains\%}()},
\code{\link{as.best}()},
\code{\link{enclose}()},
\code{\link{locf}()},
\code{\link{padded}()},
\code{\link{parens}()},
\code{\link{pool}()},
\code{\link{runhead}()},
\code{\link{text2decimal}()}
}
\concept{util}
