% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{as.best}
\alias{as.best}
\alias{as.best.data.frame}
\title{Convert to Best of Numeric or Character}
\usage{
as.best(x, ...)

\method{as.best}{data.frame}(x, prefix = "#", ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{passed arguments}

\item{prefix}{character to prepend to values in mixed numeric/character columns}
}
\value{
data.frame
}
\description{
Convert to best of numeric or character.

Convert data.frame columns to best of numeric or character.
}
\details{
Converts to numeric if doing so creates no new NA; otherwise to character.

Converts columns to numeric if doing so creates no new NA; otherwise to character. Number-like columns that are nevertheless character are prefixed by default to make this explicit when viewing only a few rows.
}
\section{Methods (by class)}{
\itemize{
\item \code{as.best(data.frame)}: data.frame method

}}
\seealso{
Other util: 
\code{\link{\%contains\%}()},
\code{\link{as.best.default}()},
\code{\link{enclose}()},
\code{\link{locf}()},
\code{\link{padded}()},
\code{\link{parens}()},
\code{\link{pool}()},
\code{\link{runhead}()},
\code{\link{text2decimal}()}

Other util: 
\code{\link{\%contains\%}()},
\code{\link{as.best.default}()},
\code{\link{enclose}()},
\code{\link{locf}()},
\code{\link{padded}()},
\code{\link{parens}()},
\code{\link{pool}()},
\code{\link{runhead}()},
\code{\link{text2decimal}()}
}
\concept{util}
\keyword{internal}
