% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{as.bootstrap.bootstrap}
\alias{as.bootstrap.bootstrap}
\title{Create Bootstrap from Bootstrap}
\usage{
\method{as.bootstrap}{bootstrap}(x, ...)
}
\arguments{
\item{x}{bootstrap}

\item{...}{ignored}
}
\value{
data.frame
}
\description{
Creates bootstrap from bootstrap.
}
\seealso{
Other as.bootstrap: 
\code{\link{as.bootstrap}()},
\code{\link{as.bootstrap.numeric}()}
}
\concept{as.bootstrap}
