% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{as.bootstrap.numeric}
\alias{as.bootstrap.numeric}
\title{Create Bootstrap from Numeric}
\usage{
\method{as.bootstrap}{numeric}(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{arguments to methods}
}
\value{
data.frame
}
\description{
Creates bootstrap from numeric.
}
\keyword{internal}
