% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical.R
\name{as.canonical.character}
\alias{as.canonical.character}
\title{Generate Canonical Names for Character}
\usage{
\method{as.canonical}{character}(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed arguments}
}
\description{
Generates canonical names for numeric by converting to parsed nmctl.
}

