% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical.R
\name{as.canonical.numeric}
\alias{as.canonical.numeric}
\title{Generate Canonical Names for Numeric}
\usage{
\method{as.canonical}{numeric}(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed arguments}
}
\description{
Generates canonical names for numeric by coercing to character.
}
