% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical.R
\name{as.canonical}
\alias{as.canonical}
\title{Generate Canonical Names}
\usage{
as.canonical(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed arguments}
}
\description{
Generates canonical names.
}
