% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{as.character.init}
\alias{as.character.init}
\title{Coerce init to character}
\usage{
\method{as.character}{init}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{dots}
}
\value{
character
}
\description{
Coerces init to character.
}
\seealso{
Other as.character: 
\code{\link{as.character.inits}()},
\code{\link{as.character.items}()},
\code{\link{as.character.model}()},
\code{\link{as.character.problem}()}
}
\concept{as.character}
\keyword{internal}
