% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.init.R
\name{as.character.initList}
\alias{as.character.initList}
\title{Coerce initList to character}
\usage{
\method{as.character}{initList}(x, ...)
}
\arguments{
\item{x}{initList}

\item{...}{dots}
}
\value{
character
}
\description{
Coerces initList to character.
}
\keyword{internal}

