% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{as.character.inits}
\alias{as.character.inits}
\title{Coerce inits to character}
\usage{
\method{as.character}{inits}(x, pretty = TRUE, sep = ";", delim = " ; ",
  widths = comwidth(x), ...)
}
\arguments{
\item{x}{inits}

\item{...}{dots}
}
\value{
character
}
\description{
Coerces inits to character.
}
\keyword{internal}
