% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{as.character.itemList}
\alias{as.character.itemList}
\title{Coerce itemList to Character}
\usage{
\method{as.character}{itemList}(x, ...)
}
\arguments{
\item{x}{itemList}

\item{...}{dots}
}
\value{
character
}
\description{
Coerces itemList to character.
}
