% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{as.character.items}
\alias{as.character.items}
\title{Coerce Items to Character}
\usage{
\method{as.character}{items}(x, ...)
}
\arguments{
\item{x}{items}

\item{...}{dots}
}
\value{
character
}
\description{
Coerces items to character.
}
