% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{as.character.model}
\alias{as.character.model}
\title{Coerce NONMEM Control Object to character}
\usage{
\method{as.character}{model}(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{dots}
}
\value{
model
}
\description{
Coerces NONMEM control stream object to character.
}
\seealso{
Other as.character: \code{\link{as.character.inits}},
  \code{\link{as.character.init}},
  \code{\link{as.character.items}}
}
\keyword{internal}
