% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.nmctl.R
\name{as.character.nmctl}
\alias{as.character.nmctl}
\title{Coerce NONMEM Control Object to character}
\usage{
\method{as.character}{nmctl}(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{dots}
}
\value{
nmctl
}
\description{
Coerces NONMEM control stream object to character.
}
\keyword{internal}

