% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{as.character.problem}
\alias{as.character.problem}
\title{Coerce Problem to Character}
\usage{
\method{as.character}{problem}(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{dots}
}
\value{
character
}
\description{
Coerces NONMEM problem statement to character.
}
\seealso{
Other as.character: 
\code{\link{as.character.inits}()},
\code{\link{as.character.init}()},
\code{\link{as.character.items}()},
\code{\link{as.character.model}()}
}
\concept{as.character}
\keyword{internal}
