% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{as.csv.character}
\alias{as.csv.character}
\title{Treat Character as CSV}
\usage{
\method{as.csv}{character}(x, as.is = TRUE, na.strings = c("", " ", "."),
  strip.white = T, ...)
}
\arguments{
\item{x}{object}

\item{as.is}{passed to read.csv}

\item{na.strings}{passed to read.csv}

\item{strip.white}{passed to read.csv}

\item{...}{passed arguments}
}
\value{
data.frame
}
\description{
Treat a character string as a csv filename.
}
\details{
If x is character, is length one, and is a path to a file, an attempt is made to read the file.
}

