% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{as.csv.data.frame}
\alias{as.csv.data.frame}
\title{Save a Data Frame as CSV}
\usage{
\method{as.csv}{data.frame}(x, file, na = ".", quote = FALSE,
  row.names = FALSE, ...)
}
\arguments{
\item{x}{object}

\item{file}{passed to write.csv}

\item{na}{passed to write.csv}

\item{quote}{passed to write.csv}

\item{row.names}{passed to write.csv}

\item{...}{passed arguments}
}
\value{
invisible data.frame (x)
}
\description{
Saves a data.frame as csv.
}

