% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{as.csv}
\alias{as.csv}
\title{Read or Write Standardized CSV
Reads or writes CSV files in a standard way.}
\usage{
as.csv(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Read or Write Standardized CSV
Reads or writes CSV files in a standard way.
}

