% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{as.definitions.character}
\alias{as.definitions.character}
\alias{definitions}
\title{Create Model Item Definitions from Character}
\usage{
\method{as.definitions}{character}(x, verbose = FALSE,
  ctlfile = modelfile(x, ...), metafile = modelpath(x, "def", ...),
  fields = c("symbol", "label", "unit"), read = length(metafile) == 1,
  write = FALSE, ...)
}
\arguments{
\item{x}{a model name (numeric or character)}

\item{verbose}{set FALSE to suppress messages}

\item{ctlfile}{path to control stream (pass length-zero argument to ignore)}

\item{metafile}{path to definitions file (pass length-zero argument to ignore)}

\item{fields}{metadata fields to read from control stream if no metafile}

\item{read}{whether to read the definitions file}

\item{write}{whether to write the definitions file}

\item{...}{passed to other functions}
}
\value{
object of class definitions, or path to metafile if write = TRUE.
}
\description{
Create Item Definitions from Model Name
}
\details{
Creates item definitions from a model name. Scavenges definitions optionally 
from the control stream and optionally from the definitions file. Optionally
writes the result to the definitions file. Always returns a data.frame with 
at least the column 'item' but possibly no rows.
}
\examples{
library(magrittr)
options(project = system.file('project/model',package='nonmemica'))
1001 \%>\% as.definitions
}
\seealso{
\code{\link{as.xml_document.character}}

\code{\link{as.bootstrap.character}}

\code{\link{as.model.character}}
}
