% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{as.definitions.modelname}
\alias{as.definitions.modelname}
\alias{definitions}
\title{Create Item Definitions from Model Name}
\usage{
\method{as.definitions}{modelname}(x, verbose = FALSE,
  opt = getOption("project"), project = if (is.null(opt)) getwd() else opt,
  rundir = file.path(project, x), ctlfile = file.path(rundir, paste0(x,
  ".ctl")), metafile = file.path(rundir, paste0(x, ".def")),
  fields = c("symbol", "label", "unit"), read = length(metafile) == 1,
  write = FALSE, ...)
}
\arguments{
\item{x}{a model name (numeric or character)}

\item{verbose}{set FALSE to suppress messages}

\item{opt}{alternative argument for setting project}

\item{project}{parent directory of model directories}

\item{rundir}{specific model directory}

\item{ctlfile}{path to control stream (pass length-zero argument to ignore)}

\item{metafile}{path to definitions file (pass length-zero argument to ignore)}

\item{fields}{metadata fields to read from control stream if no metafile}

\item{read}{whether to read the definitions file}

\item{write}{whether to write the definitions file}

\item{...}{passed to other functions}
}
\value{
object of class definitions, or path to metafile if write = TRUE.
}
\description{
Creates item definitions from a model name. Scavenges definitions optionally 
from the control stream and optionally from the definitions file. Optionally
writes the result to the definitions file. Always returns a data.frame with 
at least the column 'item' but possibly no rows.
}
\examples{
library(magrittr)
options(project = system.file('project/model',package='nonmemica'))
1001 \%>\% as.definitions
}
\seealso{
\code{\link{as.xml_document.modelname}}

\code{\link{as.bootstrap.modelname}}

\code{\link{as.nmctl.modelname}}

\code{\link{write.csv}}

\code{\link{read.csv}}
}

