% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{as.definitions}
\alias{as.definitions}
\alias{as.definitions.character}
\alias{as.definitions.definitions}
\alias{as.definitions.numeric}
\title{Harvest Model Item Definitions in Project Context}
\usage{
as.definitions(x, ...)

\method{as.definitions}{definitions}(x, ...)

\method{as.definitions}{numeric}(x, ...)

\method{as.definitions}{character}(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{arguments to methods}
}
\description{
Havests model item definitions in project context.

Creates a model item definitions from a definitions object.

Creates a model item definitions from a number.

Creates a model item definitions from a character string.
}
\details{
x can be numeric or character model name, assuming project is identified by argument or option.

Just returns the object unmodified.

Just coerces to character and calls as.definitions again.

Reclassifies x as a modelname and calls as.definitions again.
}
\section{Methods (by class)}{
\itemize{
\item \code{definitions}: definitions method

\item \code{numeric}: numeric method

\item \code{character}: character method
}}
\seealso{
\code{\link{as.definitions.modelname}}
}

