% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likebut.R
\name{as.folded.definitions}
\alias{as.folded.definitions}
\title{Coerce to Folded from Definitions
Coerces to folded from definitions  Harvests item, label, and unit for tabled items.  Stacks these in conventional folded format.}
\usage{
as.folded.definitions(x, parameters = FALSE, ...)
}
\arguments{
\item{x}{definitions}

\item{parameters}{whether to included parameter metadata}

\item{...}{passed arguments}
}
\value{
folded
}
\description{
Coerce to Folded from Definitions
Coerces to folded from definitions  Harvests item, label, and unit for tabled items.  Stacks these in conventional folded format.
}
