% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likebut.R
\name{as.folded.spec}
\alias{as.folded.spec}
\title{Coerce to Folded from Spec
Coerces to folded from spec.  Harvests column names, labels and units.  Stacks these in conventional folded format.}
\usage{
\method{as.folded}{spec}(x, ...)
}
\arguments{
\item{x}{spec}

\item{...}{passed arguments}
}
\value{
folded
}
\description{
Coerce to Folded from Spec
Coerces to folded from spec.  Harvests column names, labels and units.  Stacks these in conventional folded format.
}
