% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halfmatrix.R
\name{as.halfmatrix}
\alias{as.halfmatrix}
\title{Coerce to  Half Matrix}
\usage{
as.halfmatrix(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Coerces to half matrix. Generic, with default methods.
}
\seealso{
Other halfmatrix: 
\code{\link{as.data.frame.halfmatrix}()},
\code{\link{as.halfmatrix.default}()},
\code{\link{as.halfmatrix.halfmatrix}()},
\code{\link{as.matrix.halfmatrix}()},
\code{\link{half}()},
\code{\link{half.matrix}()},
\code{\link{is.square}()},
\code{\link{is.square.matrix}()},
\code{\link{offdiag}()},
\code{\link{offdiag.halfmatrix}()},
\code{\link{ord}()},
\code{\link{ord.halfmatrix}()},
\code{\link{ord.matrix}()},
\code{\link{print.halfmatrix}()}
}
\concept{halfmatrix}
\keyword{internal}
