% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{as.init.init}
\alias{as.init.init}
\title{Coerce init to init}
\usage{
\method{as.init}{init}(x = numeric(0), fixed = FALSE, comment = character(0), ...)
}
\arguments{
\item{x}{init}

\item{fixed}{logical}

\item{comment}{character}

\item{...}{passed arguments}
}
\value{
init
}
\description{
Coerces init to init
}
\seealso{
Other as.init: 
\code{\link{as.init.numeric}()},
\code{\link{as.init}()}
}
\concept{as.init}
