% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{as.init.numeric}
\alias{as.init.numeric}
\title{Coerces numeric to init}
\usage{
\method{as.init}{numeric}(x = numeric(0), fixed = FALSE, comment = character(0), ...)
}
\arguments{
\item{x}{object}

\item{fixed}{logical}

\item{comment}{character}

\item{...}{dots}
}
\value{
init
}
\description{
Coerces numeric to init
}
\seealso{
Other as.init: 
\code{\link{as.init}()}
}
\concept{as.init}
\keyword{internal}
