% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{as.init}
\alias{as.init}
\title{Coerce to init}
\usage{
as.init(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{dots}
}
\description{
Coerces to init
}
\seealso{
Other as.init: 
\code{\link{as.init.init}()},
\code{\link{as.init.numeric}()}
}
\concept{as.init}
\keyword{internal}
