% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.init.R
\name{as.initList}
\alias{as.initList}
\alias{as.initList.character}
\alias{as.initList.initList}
\alias{as.initList.list}
\alias{as.initList.numeric}
\title{Coerce to initList}
\usage{
as.initList(x, ...)

\method{as.initList}{list}(x, comment = character(0), ...)

\method{as.initList}{numeric}(x, fixed = FALSE, comment = character(0), ...)

\method{as.initList}{character}(x, ...)

\method{as.initList}{initList}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{dots}

\item{comment}{character}

\item{fixed}{logical}

\item{comment}{character}
}
\value{
initList

initList

initList

initList
}
\description{
Coerces to initList.

Coerces list to initlist.

Coerces numeric to initList

Coerces character to initList.

Coerces initList to initList
}
\section{Methods (by class)}{
\itemize{
\item \code{list}: list method

\item \code{numeric}: numeric method

\item \code{character}: character method

\item \code{initList}: initList method
}}
\keyword{internal}

