% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{as.inits.list}
\alias{as.inits.list}
\title{Coerce list to inits}
\usage{
\method{as.inits}{list}(x, comment = character(0), ...)
}
\arguments{
\item{x}{list}

\item{comment}{character}

\item{...}{passed arguments}
}
\value{
inits
}
\description{
Coerces list to inits.
}
\seealso{
Other as.inits: 
\code{\link{as.inits}()},
\code{\link{as.inits.numeric}()}
}
\concept{as.inits}
