% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{as.inits.numeric}
\alias{as.inits.numeric}
\title{Coerce numeric to inits}
\usage{
\method{as.inits}{numeric}(x, fixed = FALSE, comment = character(0), ...)
}
\arguments{
\item{x}{numeric}

\item{fixed}{logical}

\item{comment}{character}
}
\value{
inits
}
\description{
Coerces numeric to inits
}
\seealso{
Other as.inits: 
\code{\link{as.inits}()},
\code{\link{as.inits.list}()}
}
\concept{as.inits}
\keyword{internal}
