% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{as.inits}
\alias{as.inits}
\alias{as.inits.list}
\alias{as.inits.numeric}
\alias{as.inits.character}
\alias{as.inits.inits}
\title{Coerce to inits}
\usage{
as.inits(x, ...)

\method{as.inits}{list}(x, comment = character(0), ...)

\method{as.inits}{numeric}(x, fixed = FALSE, comment = character(0), ...)

\method{as.inits}{character}(x, ...)

\method{as.inits}{inits}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{dots}

\item{comment}{character}

\item{fixed}{logical}

\item{comment}{character}
}
\value{
inits

inits

inits

inits
}
\description{
Coerces to inits.

Coerces list to inits.

Coerces numeric to inits

Coerces character to inits.

Coerces inits to inits
}
\section{Methods (by class)}{
\itemize{
\item \code{list}: list method

\item \code{numeric}: numeric method

\item \code{character}: character method

\item \code{inits}: inits method
}}

\keyword{internal}
