% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{as.itemComments}
\alias{as.itemComments}
\alias{as.itemComments.records}
\alias{as.itemComments.model}
\alias{as.itemComments.itemList}
\alias{as.itemComments.initList}
\title{Coerce to itemComments}
\usage{
as.itemComments(x, ...)

\method{as.itemComments}{records}(x, ...)

\method{as.itemComments}{model}(x, fields = c("symbol", "unit", "label"),
  expected = character(0), na = NA_character_, tables = TRUE, ...)

\method{as.itemComments}{itemList}(x, ...)

\method{as.itemComments}{initList}(x, type, prior, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{dots}

\item{fields}{data items to scavenge from control stream comments}

\item{expected}{parameters known from NONMEM output}

\item{na}{string to use for NA values when writing default metafile}

\item{tables}{whether to include table comments}

\item{type}{item type: theta, omega, sigma (tables give itemList not initList)}

\item{prior}{number of prior items of this type (maybe imporant for numbering)}

\item{x}{model}

\item{...}{passed arguments}
}
\value{
data.frame

data.frame

data.frame

data.frame
}
\description{
Coerces to itemComments

Converts Records to itemComments

Converts model to itemComments.

Converts itemList to itemComments

Coerces to itemComments from initList.
}
\section{Methods (by class)}{
\itemize{
\item \code{records}: record method

\item \code{model}: model method

\item \code{itemList}: itemList method

\item \code{initList}: initList method
}}

\keyword{internal}
