% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.nmctl.R
\name{as.itemList.character}
\alias{as.itemList.character}
\title{Convert to itemList from Character}
\usage{
\method{as.itemList}{character}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
itemList
}
\description{
Converts to itemlist from character
}

