% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{as.items.character}
\alias{as.items.character}
\title{Convert to Items from Character}
\usage{
\method{as.items}{character}(x, ...)
}
\arguments{
\item{x}{character}

\item{...}{ignored}
}
\value{
items
}
\description{
Converts to items from character
}
\concept{as.items}
\keyword{internal}
