% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{as.list.model}
\alias{as.list.model}
\title{Coerce model to list}
\usage{
\method{as.list}{model}(x, ...)
}
\arguments{
\item{x}{model}

\item{...}{dots}
}
\value{
list
}
\description{
Coerces model to list.
}
\keyword{internal}
