% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.nmctl.R
\name{as.list.nmctl}
\alias{as.list.nmctl}
\title{Coerce nmctl to list}
\usage{
\method{as.list}{nmctl}(x, ...)
}
\arguments{
\item{x}{nmctl}

\item{...}{dots}
}
\value{
list
}
\description{
Coerces nmctl to list.
}
\keyword{internal}

