% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{as.matrices.inits}
\alias{as.matrices.inits}
\title{Coerce to Matrices from Inits}
\usage{
\method{as.matrices}{inits}(x, ...)
}
\arguments{
\item{x}{inits}

\item{...}{ignored}
}
\value{
matrices
}
\description{
Coerces to matrices from inits. Non-block inits is expanded into list of matrices.
}
\seealso{
Other as.matrices: 
\code{\link{as.matrices}()},
\code{\link{as.matrices.records}()}
}
\concept{as.matrices}
\keyword{internal}
