% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{as.matrices}
\alias{as.matrices}
\title{Coerce to List of Matrices}
\usage{
as.matrices(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{dots}
}
\description{
Coerces to list of matrices.
}
\seealso{
Other as.matrices: 
\code{\link{as.matrices.inits}()},
\code{\link{as.matrices.records}()}
}
\concept{as.matrices}
\keyword{internal}
