% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{as.matrices.records}
\alias{as.matrices.records}
\title{Coerce to List of Matrices from Records}
\usage{
\method{as.matrices}{records}(x, ...)
}
\arguments{
\item{x}{records}

\item{...}{ignored}
}
\description{
Coerces to list of matrices from Records
}
\seealso{
Other as.matrices: 
\code{\link{as.matrices}()},
\code{\link{as.matrices.inits}()}
}
\concept{as.matrices}
\keyword{internal}
