% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.nmctl.R
\name{as.matrix.initList}
\alias{as.matrix.initList}
\title{Coerce to matrix from initList}
\usage{
\method{as.matrix}{initList}(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{dots}
}
\value{
matrix
}
\description{
Coerces to matrix from initList.
}
\keyword{internal}

