% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.nmctl.R
\name{as.matrixList.initList}
\alias{as.matrixList.initList}
\title{Coerce to matrixList from initList}
\usage{
\method{as.matrixList}{initList}(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{dots}
}
\value{
matrixList
}
\description{
Coerces to matrixList from initList. Non-block initList is expanded into list of matrices.
}
\keyword{internal}

