% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.nmctl.R
\name{as.matrixList.nmctlType}
\alias{as.matrixList.nmctlType}
\title{Coerce to List of Matrices from nmctlType}
\usage{
\method{as.matrixList}{nmctlType}(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{dots}
}
\description{
Coerces to list of matrices from nmctlType
}
\keyword{internal}

