% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.nmctl.R
\name{as.matrixList}
\alias{as.matrixList}
\title{Coerce to List of Matrices}
\usage{
as.matrixList(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{dots}
}
\description{
Coerces to list of matrices.
}
\keyword{internal}

