% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{as.matrixList.records}
\alias{as.matrixList.records}
\title{Coerce to List of Matrices from Records}
\usage{
\method{as.matrixList}{records}(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{dots}
}
\description{
Coerces to list of matrices from Records
}
\keyword{internal}
