% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{as.model.numeric}
\alias{as.model.numeric}
\title{Coerce to Model from Numeric}
\usage{
\method{as.model}{numeric}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{dots}
}
\description{
Coerces to model from numeric by coercing to character.
}
\seealso{
Other as.model: 
\code{\link{[.model}()},
\code{\link{[[.model}()},
\code{\link{as.model}()},
\code{\link{read.model}()},
\code{\link{write.model}()}
}
\concept{as.model}
\keyword{internal}
