% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{as.model}
\alias{as.model}
\title{Coerce to NONMEM Control Object}
\usage{
as.model(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed arguments}
}
\value{
model
}
\description{
Coerces to NONMEM control stream object.
}
\seealso{
Other as.model: 
\code{\link{[.model}()},
\code{\link{[[.model}()},
\code{\link{as.model.character}()},
\code{\link{as.model.numeric}()},
\code{\link{read.model}()},
\code{\link{write.model}()}
}
\concept{as.model}
\keyword{internal}
