% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical.R
\name{as.nonmem.character}
\alias{as.nonmem.character}
\title{Generate NONMEM-style Names for Character}
\usage{
\method{as.nonmem}{character}(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed arguments}
}
\description{
Generates NONMEM-style names for numeric by converting to parsed model.
}
