% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical.R
\name{as.nonmem.nmctl}
\alias{as.nonmem.nmctl}
\title{Generate NONMEM-style Names for nmctl}
\usage{
\method{as.nonmem}{nmctl}(x, ...)
}
\arguments{
\item{x}{a model designator}

\item{...}{passed arguments}
}
\value{
nonmem (character)
}
\description{
Generates NONMEM-style names for parameters declared in a NONMEM control stream object. PsN uses NONMEM-style names, substituting a comment, if any: everything after the first semicolon, up to the second semicolon if present, without leading/trailing spaces/tabs.
}
\seealso{
as.nmctl
}
