% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical.R
\name{as.nonmem.numeric}
\alias{as.nonmem.numeric}
\title{Generate NONMEM-style Names for Numeric}
\usage{
\method{as.nonmem}{numeric}(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed arguments}
}
\description{
Generates NONMEM-style names for numeric by coercing to character.
}
