% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical.R
\name{as.nonmem}
\alias{as.nonmem}
\title{Generate NONMEM-style Names}
\usage{
as.nonmem(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed arguments}
}
\description{
Generates NONMEM-style names.
}
