% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{as.omega.model}
\alias{as.omega.model}
\title{Extract Omegas from Model}
\usage{
\method{as.omega}{model}(x, ...)
}
\arguments{
\item{x}{model}

\item{...}{passed arguments}
}
\value{
omega (subset of model)
}
\description{
Extracts omegas from model.
}
