% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{as.omega}
\alias{as.omega}
\title{Extract Omegas}
\usage{
as.omega(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Extracts omegas.
}
\seealso{
Other as.omega: \code{\link{as.omega.model}}
}
\keyword{internal}
