% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partab.R
\name{as.partab}
\alias{as.partab}
\alias{as.partab.character}
\alias{as.partab.numeric}
\alias{as.partab.partab}
\title{Create Model Parameter Table in Project Context}
\usage{
as.partab(x, ...)

\method{as.partab}{partab}(x, ...)

\method{as.partab}{numeric}(x, ...)

\method{as.partab}{character}(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{arguments to methods}
}
\description{
Creates a model parameter table in project context.

Creates a model parameter table from a partab object.

Creates a model parameter table from a number.

Creates a model parameter table from a character string.
}
\details{
x can be numeric or character model name, assuming project is identified by argument or option.

Just returns the object unmodified.

Just coerces to character and calls as.partab again.

Reclassifies x as a modelname and calls as.partab again.
}
\section{Methods (by class)}{
\itemize{
\item \code{partab}: partab method

\item \code{numeric}: numeric method

\item \code{character}: character method
}}
\seealso{
\code{\link{as.partab.modelname}}
}

