% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likebut.R
\name{as.pharmval.character}
\alias{as.pharmval.character}
\title{Coerce Character to Pharmval}
\usage{
\method{as.pharmval}{character}(x, ..., pvlfile, ext = "pvl")
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}

\item{pvlfile}{if not supplied, guessed to be model's datafile except with 
the value of ext in place of 'csv'.}

\item{ext}{default extention for pharmval version of model datafile.}
}
\value{
meta data.frame
}
\description{
Coerces to pharmval by treating character as modelname.  Combines the metasuperset
(model ouptut in meta format) with the meta version of the model data.file.
}
\seealso{
metasuperset
}

