% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likebut.R
\name{as.pharmval.numeric}
\alias{as.pharmval.numeric}
\title{Coerce Numeric to Pharmval}
\usage{
\method{as.pharmval}{numeric}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Coerces numeric to pharmval by coercing first to character.
}

