% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical.R
\name{as.psn.character}
\alias{as.psn.character}
\title{Generate PsN-style Names for Character}
\usage{
\method{as.psn}{character}(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed arguments}
}
\description{
Generates PsN-style names for numeric by converting to parsed nmctl.
}

