% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical.R
\name{as.psn.numeric}
\alias{as.psn.numeric}
\title{Generate PsN-style Names for Numeric}
\usage{
\method{as.psn}{numeric}(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed arguments}
}
\description{
Generates PsN-style names for numeric by coercing to character.
}

