% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{as.sigma.model}
\alias{as.sigma.model}
\title{Extract Sigmas from Model}
\usage{
\method{as.sigma}{model}(x, ...)
}
\arguments{
\item{x}{model}

\item{...}{passed arguments}
}
\value{
sigma (subset of model)
}
\description{
Extracts sigmas from model.
}
\concept{as.sigma}
\keyword{internal}
