% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superset.R
\name{as.superset}
\alias{as.superset}
\title{Coerce to superset}
\usage{
as.superset(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Coerces to superset.
}
\details{
Generic, with methods for numeric and character.
}

