% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{as.tab.model}
\alias{as.tab.model}
\title{Extract Tables from Model}
\usage{
\method{as.tab}{model}(x, ...)
}
\arguments{
\item{x}{model}

\item{...}{passed arguments}
}
\value{
tab (subset of model)
}
\description{
Extracts tables from model.
}
\keyword{internal}
