% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{as.tab}
\alias{as.tab}
\title{Extract Tables}
\usage{
as.tab(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Extracts tables.
}
\seealso{
Other as.tab: \code{\link{as.tab.model}}
}
\keyword{internal}
