% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{as.theta.model}
\alias{as.theta.model}
\title{Extract Thetas from Model}
\usage{
\method{as.theta}{model}(x, ...)
}
\arguments{
\item{x}{model}

\item{...}{passed arguments}
}
\value{
theta (subset of model)
}
\description{
Extracts thetas from model.
}
\seealso{
Other as.theta: 
\code{\link{as.theta}()}
}
\concept{as.theta}
\keyword{internal}
