% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{as.theta}
\alias{as.theta}
\title{Extract Thetas}
\usage{
as.theta(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Extracts thetas.
}
\seealso{
Other as.theta: 
\code{\link{as.theta.model}()}
}
\concept{as.theta}
\keyword{internal}
