% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpath.R
\name{as.xml_document.character}
\alias{as.xml_document.character}
\title{Create xml_document From Character}
\usage{
\method{as.xml_document}{character}(x, strip.namespace = TRUE, ...)
}
\arguments{
\item{x}{file path or run name}

\item{strip.namespace}{whether to strip e.g. nm: from xml elements}

\item{...}{passed to modelpath()}
}
\value{
xml_document
}
\description{
Creates an xml_document from character (modelname or filepath).
}
\seealso{
Other xpath: 
\code{\link{as.xml_document.numeric}()},
\code{\link{as.xml_document}()},
\code{\link{xpath.default}()},
\code{\link{xpath.xml_document}()},
\code{\link{xpath}()}
}
\concept{xpath}
