% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likebut.R
\name{but}
\alias{but}
\title{Identify a Distinctive Feature}
\usage{
but(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Identifies a distinctive feature of an object.
}
\keyword{internal}
