% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{$.init}
\alias{$.init}
\title{Select init element}
\usage{
\method{$}{init}(x, name)
}
\arguments{
\item{x}{init}

\item{name}{character}
}
\value{
numeric
}
\description{
Selects init element.
}
\seealso{
Other init: \code{\link{$<-.init}}
}
\keyword{internal}
