% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{comments.inits}
\alias{comments.inits}
\title{Extract Comments from Inits}
\usage{
\method{comments}{inits}(x, type, prior, ...)
}
\arguments{
\item{x}{inits}

\item{type}{item type: theta, omega, sigma (tables give items not inits)}

\item{prior}{number of prior items of this type (maybe imporant for numbering)}

\item{...}{ignored}
}
\value{
data.frame
}
\description{
Extracts comments from inits.
}
\seealso{
Other comments: 
\code{\link{comments.items}()},
\code{\link{comments.model}()},
\code{\link{comments}()}
}
\concept{comments}
