% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{comments.items}
\alias{comments.items}
\title{Extract Comments from Items}
\usage{
\method{comments}{items}(x, ...)
}
\arguments{
\item{x}{items}

\item{...}{ignored}
}
\value{
data.frame
}
\description{
Extracts comments from items.
}
\seealso{
Other comments: 
\code{\link{comments}()},
\code{\link{comments.inits}()},
\code{\link{comments.model}()}
}
\concept{comments}
