% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{comments}
\alias{comments}
\alias{comments.records}
\title{Extract Comments}
\usage{
comments(x, ...)

\method{comments}{records}(x, ...)
}
\arguments{
\item{x}{records}

\item{...}{ignored}
}
\value{
data.frame
}
\description{
Extracts comments.

Extracts comments from records.
}
\section{Methods (by class)}{
\itemize{
\item \code{comments(records)}: record method

}}
\seealso{
Other comments: 
\code{\link{comments.inits}()},
\code{\link{comments.items}()},
\code{\link{comments.model}()}

Other comments: 
\code{\link{comments.inits}()},
\code{\link{comments.items}()},
\code{\link{comments.model}()}
}
\concept{comments}
\keyword{internal}
