% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{comwidth.character}
\alias{comwidth.character}
\title{Calculate Comment Widths for Character}
\usage{
\method{comwidth}{character}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Calculates comment widths for character.
}
\seealso{
Other comwidth: 
\code{\link{comwidth.inits}()},
\code{\link{comwidth.items}()},
\code{\link{comwidth}()}
}
\concept{comwidth}
\keyword{internal}
