% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{comwidth.items}
\alias{comwidth.items}
\title{Calculate Comment Widths for Items}
\usage{
\method{comwidth}{items}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Calculates comment widths for items
}
\keyword{internal}
