% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{comwidth}
\alias{comwidth}
\title{Calculate Comment Widths
Calculates comment widths.}
\usage{
comwidth(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Calculate Comment Widths
Calculates comment widths.
}
\keyword{internal}
