% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{comwidthOne.character}
\alias{comwidthOne.character}
\title{Calculate Comment Widths for One Element of Character}
\usage{
\method{comwidthOne}{character}(x, split = ";", ...)
}
\arguments{
\item{x}{object}

\item{split}{comment separator}

\item{...}{passed arguments}
}
\description{
Calculates Comment Widths for one element of character
}
\seealso{
Other comwidthOne: 
\code{\link{comwidthOne}()}
}
\concept{comwidthOne}
\keyword{internal}
