% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{comwidthOne}
\alias{comwidthOne}
\title{Calculate Comment Widths for One Element}
\usage{
comwidthOne(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Calculates comment widths for one element.
}
\keyword{internal}
