% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{contains}
\alias{contains}
\title{Check Whether Text Contains Pattern}
\usage{
contains(pattern, text, ...)
}
\arguments{
\item{pattern}{regular expression}

\item{text}{character vector to check}

\item{...}{arguments to methods}
}
\value{
logical
}
\description{
Checks whether text contains pattern.
}
\examples{
contains('a',letters)
}
\seealso{
\code{\link{\%contains\%}}
}
