% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{datafile.numeric}
\alias{datafile.numeric}
\title{Identify Datafile from Numeric}
\usage{
\method{datafile}{numeric}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{dots}
}
\description{
Identifies datafile from numeric.
}
\keyword{internal}
