% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{datafile}
\alias{datafile}
\title{Identify Datafile}
\usage{
datafile(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{dots}
}
\description{
Identifies datafile.
}
\seealso{
Other path: 
\code{\link{datafile.character}()},
\code{\link{datafile.numeric}()},
\code{\link{modeldir}()},
\code{\link{modelfile}()},
\code{\link{modelpath}()},
\code{\link{modelpath.character}()},
\code{\link{modelpath.numeric}()},
\code{\link{psn_options}()},
\code{\link{specfile}()},
\code{\link{specfile.character}()},
\code{\link{specfile.numeric}()}
}
\concept{path}
\keyword{internal}
