% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{definitions.definitions}
\alias{definitions.definitions}
\title{Create Model Item Definitions from definitions}
\usage{
\method{definitions}{definitions}(x, ...)
}
\arguments{
\item{x}{definitions}

\item{...}{ignored}
}
\description{
Creates a model item definitions from a definitions object.
}
\details{
Just returns the object unmodified.
}
\seealso{
Other definitions: 
\code{\link{definitions.numeric}()},
\code{\link{definitions}()}
}
\concept{definitions}
