% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{definitions.numeric}
\alias{definitions.numeric}
\title{Create Model Item Definitions from Number.}
\usage{
\method{definitions}{numeric}(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{arguments to methods}
}
\description{
Creates a model item definitions from a number.
}
\details{
Just coerces to character and calls definitions again.
}
\seealso{
Other definitions: 
\code{\link{definitions}()}
}
\concept{definitions}
\keyword{internal}
