% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{definitions}
\alias{definitions}
\alias{definitions.definitions}
\alias{definitions.numeric}
\title{Harvest Model Item Definitions}
\usage{
definitions(x, ...)

\method{definitions}{definitions}(x, ...)

\method{definitions}{numeric}(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{arguments to methods}
}
\description{
Havests model item definitions.

Creates a model item definitions from a definitions object.

Creates a model item definitions from a number.
}
\details{
x can be numeric or character model name, assuming project is identified by argument or option.

Just returns the object unmodified.

Just coerces to character and calls definitions again.
}
\section{Methods (by class)}{
\itemize{
\item \code{definitions}: definitions method

\item \code{numeric}: numeric method
}}

\seealso{
\code{\link{definitions.character}}
}
