% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likebut.R
\name{depends.default}
\alias{depends.default}
\title{Identify Model Dependencies}
\usage{
\method{depends}{default}(x, opt = getOption("project"), project = if
  (is.null(opt)) getwd() else opt, ...)
}
\arguments{
\item{x}{object}

\item{opt}{alternative specification of project directory}

\item{project}{direct specification of project directory}

\item{...}{passed arguments}
}
\value{
character
}
\description{
Identify those models in the lineage of models in x.
}

