% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likebut.R
\name{depends}
\alias{depends}
\title{Identify What Something Depends On}
\usage{
depends(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Identfies that on which something depends.
}

