% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{enclose}
\alias{enclose}
\title{Enclose in Arbitrary Characters}
\usage{
enclose(x, open, close, ...)
}
\arguments{
\item{x}{vector}

\item{open}{open string}

\item{close}{close string}

\item{...}{dots}
}
\value{
character
}
\description{
Enclose in arbitrary characters
}
\keyword{internal}
