% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likebut.R
\name{errors.character}
\alias{errors.character}
\title{Get Errors for Character.}
\usage{
\method{errors}{character}(x, project = if (is.null(opt)) getwd() else opt,
  opt = getOption("project"), rundir = file.path(project, x),
  xmlfile = file.path(rundir, paste0(x, ".xml")),
  ctlfile = file.path(rundir, paste0(x, ".ctl")), strip.namespace = TRUE,
  digits = 3, ...)
}
\arguments{
\item{x}{character (modelname)}

\item{project}{parent directory of model directories}

\item{opt}{alternative argument for setting project}

\item{rundir}{specific model directory}

\item{xmlfile}{path to xml file}

\item{ctlfile}{path to control stream}

\item{strip.namespace}{whether to strip e.g. nm: from xml elements for easier xpath syntax}

\item{digits}{passed to signif}

\item{...}{dots}
}
\value{
numeric
}
\description{
Gets model asymptotic standard errors in canonical order, treating character as model names. 
See \code{\link{parameters}} for a less formal interface.
}
\seealso{
as.canonical errors
}

