% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likebut.R
\name{errors.numeric}
\alias{errors.numeric}
\title{Get Errors for Numeric}
\usage{
\method{errors}{numeric}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Gets errors for numeric by coercing to character.
}

