% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likebut.R
\name{estimates.numeric}
\alias{estimates.numeric}
\title{Get Estimates for Numeric}
\usage{
\method{estimates}{numeric}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Gets estimates for numeric by coercing to character.
}
\seealso{
Other estimates: \code{\link{estimates.character}},
  \code{\link{estimates}}
}
\keyword{internal}
