% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likebut.R
\name{estimates}
\alias{estimates}
\title{Get Estimates}
\usage{
estimates(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Gets estimates
}
\seealso{
Other estimates: 
\code{\link{estimates.character}()},
\code{\link{estimates.numeric}()}
}
\concept{estimates}
\keyword{internal}
