% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpath.R
\name{filter_estimation}
\alias{filter_estimation}
\title{Filter XML for Relevant Estimation Step}
\usage{
filter_estimation(x, estimation = integer(0), ...)
}
\arguments{
\item{x}{file path}

\item{estimation}{integer indicating which estimation step to select, if multiple. Defaults to last.}

\item{...}{passed arguments}
}
\value{
xml_document
}
\description{
Filters XML for relevant estimation step.  Defaults to last.
}
\seealso{
Other xpath: 
\code{\link{as.xml_document}()},
\code{\link{as.xml_document.character}()},
\code{\link{as.xml_document.numeric}()},
\code{\link{readxml}()},
\code{\link{xpath}()},
\code{\link{xpath.default}()},
\code{\link{xpath.xml_document}()}
}
\concept{xpath}
