% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{fixed<-.init}
\alias{fixed<-.init}
\title{Set fixed init value}
\usage{
\method{fixed}{init}(x) <- value
}
\arguments{
\item{x}{object}

\item{value}{value to}
}
\value{
init
}
\description{
Sets fixed init value.
}
\seealso{
Other fixed: 
\code{\link{fixed.model}()},
\code{\link{fixed<-.inits}()},
\code{\link{fixed<-.model}()},
\code{\link{fixed<-}()},
\code{\link{fixed}()}
}
\concept{fixed}
\keyword{internal}
