% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.init.R
\name{fixed<-.initList}
\alias{fixed<-.initList}
\title{Set fixed attribute of initList}
\usage{
\method{fixed}{initList}(x) <- value
}
\arguments{
\item{x}{initList}

\item{value}{logical}
}
\value{
initList
}
\description{
Sets fixed attribute of initList.
}
\keyword{internal}

