% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{fixed<-.inits}
\alias{fixed<-.inits}
\title{Set fixed attribute of inits}
\usage{
\method{fixed}{inits}(x) <- value
}
\arguments{
\item{x}{inits}

\item{value}{logical}
}
\value{
inits
}
\description{
Sets fixed attribute of inits.
}
\seealso{
Other fixed: 
\code{\link{fixed.model}()},
\code{\link{fixed<-.init}()},
\code{\link{fixed<-.model}()},
\code{\link{fixed<-}()},
\code{\link{fixed}()}
}
\concept{fixed}
\keyword{internal}
