% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getset.R
\name{fixed<-.model}
\alias{fixed<-.model}
\title{Set fixed attribute of model}
\usage{
\method{fixed}{model}(x) <- value
}
\arguments{
\item{x}{model}

\item{value}{logical}
}
\value{
model
}
\description{
Sets fixed attribute of model.
}
\seealso{
Other fixed: 
\code{\link{fixed.model}()},
\code{\link{fixed<-.inits}()},
\code{\link{fixed<-.init}()},
\code{\link{fixed<-}()},
\code{\link{fixed}()}
}
\concept{fixed}
\keyword{internal}
