% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getset.R
\name{fixed<-.nmctl}
\alias{fixed<-.nmctl}
\title{Set fixed attribute of nmctl}
\usage{
\method{fixed}{nmctl}(x) <- value
}
\arguments{
\item{x}{nmctl}

\item{value}{logical}
}
\value{
nmctl
}
\description{
Sets fixed attribute of nmctl.
}
\keyword{internal}
