% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.init.R
\name{fixed<-}
\alias{fixed<-}
\title{Set value of fixed attribute}
\usage{
fixed(x) <- value
}
\arguments{
\item{x}{object}

\item{value}{value to set}
}
\description{
Sets value of fixed attribute.
}
\keyword{internal}

