% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{fixed.init}
\alias{fixed.init}
\title{Check if init is fixed}
\usage{
\method{fixed}{init}(x, ...)
}
\arguments{
\item{x}{init}

\item{...}{ignored}
}
\value{
logical
}
\description{
Checks if init is fixed.
}
\seealso{
Other fixed: 
\code{\link{fixed}()},
\code{\link{fixed.model}()},
\code{\link{fixed<-}()},
\code{\link{fixed<-.init}()},
\code{\link{fixed<-.inits}()},
\code{\link{fixed<-.model}()}
}
\concept{fixed}
