% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getset.R
\name{fixed.nmctl}
\alias{fixed.nmctl}
\title{Check If nmctl is Fixed}
\usage{
\method{fixed}{nmctl}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{dots}
}
\value{
logical
}
\description{
Checks if nmctl is fixed.  Returns a logical vector with element for each init, in canonical order.
}
