% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superset.R
\name{fold.character}
\alias{fold.character}
\title{Fold Character}
\usage{
\method{fold}{character}(x, ..., meta, subset)
}
\arguments{
\item{x}{numeric}

\item{...}{unquoted grouping variables, forwarded to \code{\link{fold_.character}}}

\item{meta}{pre-folded metadata}

\item{subset}{length-one character: a condition for filtering results, e.g. 'EVID #' @export}
}
\description{
Folds character, treating \code{x} as a model name.
}
\keyword{internal}
