% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superset.R
\name{fold.numeric}
\alias{fold.numeric}
\title{Fold Numeric}
\usage{
\method{fold}{numeric}(x, ...)
}
\arguments{
\item{x}{numeric}

\item{...}{passed to \code{fold}}
}
\description{
Folds numeric, coercing to character
}
\seealso{
fold.character
}
\keyword{internal}
