% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superset.R
\name{fold_.character}
\alias{fold_.character}
\title{Fold Character with Standard Evaluation}
\usage{
\method{fold_}{character}(x, ..., group_by, meta = match.fun("meta")(x, ...),
  subset)
}
\arguments{
\item{x}{character model name}

\item{...}{passed arguments}

\item{group_by}{columns to group by, i.e. key for the model's data (named values will be dropped)}

\item{meta}{pre-folded metadata}

\item{subset}{length-one character: a condition for filtering results, e.g. 'EVID}
}
\value{
folded
}
\description{
Folds character (model name) using standard evaluation.
}
