% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superset.R
\name{fold_character}
\alias{fold_character}
\title{Fold Character with Standard Evaluation}
\usage{
fold_character(x, groups, meta = match.fun("meta")(x, ...), simplify = TRUE,
  sort = TRUE, subset, ...)
}
\arguments{
\item{x}{character model name}

\item{groups}{columns to group by, i.e. key for the model's data (named values will be dropped)}

\item{meta}{pre-folded metadata}

\item{simplify}{whether to simplify the result}

\item{sort}{whether to sort the result}

\item{subset}{length-one character: a condition for filtering results, e.g. 'EVID == 1'}

\item{...}{passed arguments}
}
\value{
folded
}
\description{
Folds character (model name) using standard evaluation.
}
