% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{format.init}
\alias{format.init}
\title{Format init}
\usage{
\method{format}{init}(x, ...)
}
\arguments{
\item{x}{init}

\item{...}{dots}
}
\value{
character
}
\description{
Formats init.
}
\seealso{
Other format: 
\code{\link{format.inits}()},
\code{\link{format.items}()},
\code{\link{format.model}()}
}
\concept{format}
\keyword{internal}
