% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.init.R
\name{format.initList}
\alias{format.initList}
\title{Format initList}
\usage{
\method{format}{initList}(x, ...)
}
\arguments{
\item{x}{initList}

\item{...}{dots}
}
\value{
character
}
\description{
Formats initList.
}
\keyword{internal}

