% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{format.itemList}
\alias{format.itemList}
\title{Format itemList}
\usage{
\method{format}{itemList}(x, ...)
}
\arguments{
\item{x}{itemList}

\item{...}{dots}
}
\value{
character
}
\description{
Formats itemList.
}
\keyword{internal}
