% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{format.items}
\alias{format.items}
\title{Format Items}
\usage{
\method{format}{items}(x, ...)
}
\arguments{
\item{x}{items}

\item{...}{passed arguments}
}
\value{
character
}
\description{
Formats items.
}
\seealso{
Other format: 
\code{\link{format.inits}()},
\code{\link{format.init}()},
\code{\link{format.model}()}
}
\concept{format}
\keyword{internal}
