% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{format.model}
\alias{format.model}
\title{Format model}
\usage{
\method{format}{model}(x, ...)
}
\arguments{
\item{x}{model}

\item{...}{passed arguments}
}
\value{
character
}
\description{
Format model.
}
\details{
Coerces to character.
}
\seealso{
Other format: 
\code{\link{format.init}()},
\code{\link{format.inits}()},
\code{\link{format.items}()}
}
\concept{format}
\keyword{internal}
