% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.nmctl.R
\name{format.nmctl}
\alias{format.nmctl}
\title{Format nmctl}
\usage{
\method{format}{nmctl}(x, ...)
}
\arguments{
\item{x}{nmctl}

\item{...}{dots}
}
\value{
character
}
\description{
Format nmctl.
}
\details{
Coerces to character.
}
\keyword{internal}

