% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superset.R
\name{generalize}
\alias{generalize}
\title{Generalize a Nonmissing Value}
\usage{
generalize(x, ...)
}
\arguments{
\item{x}{vector}

\item{...}{ignored}
}
\description{
#Generalize a nonmissing value.  If there is only one such among zero or more NA, impute that value for all NA.
}
