% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halfmatrix.R
\name{half}
\alias{half}
\title{Find Half of Something}
\usage{
half(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Finds half of something.  Generic, with method for matrix.
}
\seealso{
Other halfmatrix: 
\code{\link{as.data.frame.halfmatrix}()},
\code{\link{as.halfmatrix.default}()},
\code{\link{as.halfmatrix.halfmatrix}()},
\code{\link{as.halfmatrix}()},
\code{\link{as.matrix.halfmatrix}()},
\code{\link{half.matrix}()},
\code{\link{is.square.matrix}()},
\code{\link{is.square}()},
\code{\link{offdiag.halfmatrix}()},
\code{\link{offdiag}()},
\code{\link{ord.halfmatrix}()},
\code{\link{ord.matrix}()},
\code{\link{ord}()},
\code{\link{print.halfmatrix}()}
}
\concept{halfmatrix}
\keyword{internal}
