% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{initDex.model}
\alias{initDex.model}
\title{Identify Indices of Initial Estimates in model}
\usage{
\method{initDex}{model}(x, ...)
}
\arguments{
\item{x}{model}

\item{...}{dots}
}
\value{
integer
}
\description{
Identifies record indices of initial estimates for an object of class model. If model has not been parsed, the result is integer(0).  Otherwise, the result is the record numbers for the canonical order of all init objects among theta, omega, and sigma element types, regardless of the number and order of such types. If a block(2) omega is specified between two thetas and one sigma follows, the results could be c(6L, 8L, 7L, 7L, 7L, 9L).
}
\seealso{
Other initDex: 
\code{\link{initDex}()}
}
\concept{initDex}
\keyword{internal}
