% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{initDex}
\alias{initDex}
\title{Identify Indices of Initial Estimates}
\usage{
initDex(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{dots}
}
\description{
Identifies indices of initial Estimates.
}
\seealso{
Other initDex: \code{\link{initDex.model}}
}
\keyword{internal}
