% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getset.R
\name{initial<-.model}
\alias{initial<-.model}
\title{Set Upper Bounds for Model Initial Estimates}
\usage{
\method{initial}{model}(x) <- value
}
\arguments{
\item{x}{model}

\item{value}{numeric}
}
\description{
Sets upper bounds for model initial estimates.
}
\seealso{
Other initial: 
\code{\link{initial.model}()},
\code{\link{initial<-}()},
\code{\link{initial}()}
}
\concept{initial}
