% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getset.R
\name{initial<-}
\alias{initial<-}
\title{Set Initial Value}
\usage{
initial(x) <- value
}
\arguments{
\item{x}{object of dispatch}

\item{value}{right hand side}
}
\description{
Sets initial Value.
}
\seealso{
Other initial: 
\code{\link{initial.model}()},
\code{\link{initial<-.model}()},
\code{\link{initial}()}
}
\concept{initial}
\keyword{internal}
