% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getset.R
\name{initial.model}
\alias{initial.model}
\title{Get Model Initial Estimates}
\usage{
\method{initial}{model}(x, ...)
}
\arguments{
\item{x}{model}

\item{...}{dots}
}
\description{
Gets model initial estimates.
}
