% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getset.R
\name{initial.nmctl}
\alias{initial.nmctl}
\title{Get Model Initial Estimates}
\usage{
\method{initial}{nmctl}(x, ...)
}
\arguments{
\item{x}{nmctl}

\item{...}{dots}
}
\description{
Gets model initial estimates.
}

