% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getset.R
\name{initial}
\alias{initial}
\title{Get Initial Value}
\usage{
initial(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{dots}
}
\description{
Gets initial Value.
}
\seealso{
Other initial: 
\code{\link{initial.model}()},
\code{\link{initial<-.model}()},
\code{\link{initial<-}()}
}
\concept{initial}
\keyword{internal}
