% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{initSubscripts.model}
\alias{initSubscripts.model}
\title{Identify Subscripts of Initial Estimates in model}
\usage{
\method{initSubscripts}{model}(x, ...)
}
\arguments{
\item{x}{model}

\item{...}{dots}
}
\value{
integer
}
\description{
Identifies subscripts of record indices of initial estimates for an object of class model. If model has not been parsed, the result is integer(0).  Otherwise, the result is the element number for each init object within each inits in x (canonical order).
}
\seealso{
Other initSubscripts: 
\code{\link{initSubscripts}()}
}
\concept{initSubscripts}
\keyword{internal}
