% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.nmctl.R
\name{initSubscripts.nmctl}
\alias{initSubscripts.nmctl}
\title{Identify Subscripts of Initial Estimates in nmctl}
\usage{
\method{initSubscripts}{nmctl}(x, ...)
}
\arguments{
\item{x}{nmctl}

\item{...}{dots}
}
\value{
integer
}
\description{
Identifies subscripts of record indices of initial estimates for an object of class nmctl. If nmctl has not been parsed, the result is integer(0).  Otherwise, the result is the element number for each init object within each initList in x (canonical order).
}
\keyword{internal}
