% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is.defined}
\alias{is.defined}
\title{Check Whether elements are Defined.}
\usage{
is.defined(x)
}
\arguments{
\item{x}{vector}
}
\value{
logical
}
\description{
Checks whether elements are defined.  Inverse of is.na().
}
\examples{
is.defined(c(1,NA,2))
}
