% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likebut.R
\name{like.default}
\alias{like.default}
\title{Identify the Relevant Reference Model}
\usage{
\method{like}{default}(x, opt = getOption("project"), project = if
  (is.null(opt)) getwd() else opt, ...)
}
\arguments{
\item{x}{object}

\item{opt}{alternative specification of project directory}

\item{project}{direct specification of project directory}

\item{...}{passed arguments}
}
\value{
character
}
\description{
Identifies the relevant reference model, i.e. the 'parent' of x.
}
