% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getset.R
\name{lower<-.model}
\alias{lower<-.model}
\title{Set Lower Bounds for Model Initial Estimates}
\usage{
\method{lower}{model}(x) <- value
}
\arguments{
\item{x}{model}

\item{value}{numeric}
}
\description{
Sets lower bounds for model initial estimates.
}
\seealso{
Other lower: 
\code{\link{lower}()},
\code{\link{lower.model}()},
\code{\link{lower<-}()}
}
\concept{lower}
