% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getset.R
\name{lower<-.nmctl}
\alias{lower<-.nmctl}
\title{Set Lower Bounds for Model Initial Estimates}
\usage{
\method{lower}{nmctl}(x) <- value
}
\arguments{
\item{x}{nmctl}

\item{value}{numeric}
}
\description{
Sets lower bounds for model initial estimates.
}

