% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getset.R
\name{lower<-}
\alias{lower<-}
\title{Set Lower Value}
\usage{
lower(x, ...) <- value
}
\arguments{
\item{x}{object of dispatch}

\item{value}{right hand side}

\item{...}{dots}
}
\description{
Sets lower Value.
}

