% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getset.R
\name{lower.nmctl}
\alias{lower.nmctl}
\title{Get Lower Bounds for Model Initial Estimates}
\usage{
\method{lower}{nmctl}(x, ...)
}
\arguments{
\item{x}{nmctl}

\item{...}{dots}
}
\description{
Gets lower bounds for model initial estimates.
}

