% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{maxWidths.list}
\alias{maxWidths.list}
\title{Calculate Maximum Widths for List}
\usage{
\method{maxWidths}{list}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Calculates maximum widths for list
}
\keyword{internal}
