% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{maxWidths}
\alias{maxWidths}
\title{Calculate Maximum Widths}
\usage{
maxWidths(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Calculates maximum widths
}
\seealso{
Other maxWidths: 
\code{\link{maxWidths.list}()}
}
\concept{maxWidths}
\keyword{internal}
