% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superset.R
\name{meta.character}
\alias{meta.character}
\title{Get Metadata for Character}
\usage{
\method{meta}{character}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
folded
}
\description{
Gets metadata for character, treating it as a model name.
}
