% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superset.R
\name{meta.numeric}
\alias{meta.numeric}
\title{Get Metadata for Numeric}
\usage{
\method{meta}{numeric}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Gets metadata for numeric by coercing to character.
}
\keyword{internal}
