% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superset.R
\name{metaplot.numeric}
\alias{metaplot.numeric}
\title{Metaplot Numeric}
\usage{
metaplot.numeric(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Plots numeric by coercing first to character.
}
