% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superset.R
\name{metaplot_.character}
\alias{metaplot_.character}
\title{Metaplot Character, Standard Evaluation}
\usage{
\method{metaplot_}{character}(x, groups, meta = NULL, simplify = TRUE,
  sort = TRUE, subset, var, ...)
}
\arguments{
\item{x}{object}

\item{groups}{columns by which to group the dataset, passed to fold_}

\item{meta}{pre-folded metadata; meta(x) by default}

\item{simplify}{whether to simplify the result, passed to fold_}

\item{sort}{whether to sort the result, passed to fold_}

\item{subset}{a condition for filtering data, passed to fold_}

\item{var}{variables to plot}

\item{...}{passed arguments}
}
\description{
Plots character by treating as model name.  A dataset
is constructed by combining the meta version of the model input with a
meta version of the model output and calling metaplot with the result.
}
\seealso{
\code{fold}
}
