% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likebut.R
\name{metasuperset}
\alias{metasuperset}
\title{Retrieve Model Outputs in Meta Format}
\usage{
metasuperset(x, opt = getOption("project"), project = if (is.null(opt))
  getwd() else opt, group_by = c("USUBJID", "DATETIME"),
  meta = as.definitions(x, ...), ...)
}
\arguments{
\item{x}{model name}

\item{opt}{alternative specification of project directory}

\item{project}{direct specification of project directory}

\item{group_by}{vector of key column names in superset}

\item{meta}{data.frame with item, symbol, unit, label}

\item{...}{passed arguments}
}
\value{
meta
}
\description{
Retrieves model outputs in meta format. Inputs should have EVID.
}

