% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{modeldir}
\alias{modeldir}
\title{Identify the Directory for a Model}
\usage{
modeldir(x, ext, ...)
}
\arguments{
\item{x}{the model name}

\item{ext}{model file extension}

\item{...}{passed arguments}
}
\value{
character
}
\description{
Identifies the directory used by a model.
}
\examples{
library(magrittr)
options(project = system.file('project/model',package='nonmemica'))
1001 \%>\% modeldir
}
