% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{modelpath.character}
\alias{modelpath.character}
\title{Resolve A Path to a Model-related File for Character}
\usage{
\method{modelpath}{character}(x, ext = NULL, project = getOption("project",
  getwd()), nested = getOption("nested", TRUE), ...)
}
\arguments{
\item{x}{object}

\item{ext}{file extension, no leading dot}

\item{project}{project directory}

\item{nested}{whether model files are nested in eponymous directories}

\item{...}{passed arguments}
}
\value{
character
}
\description{
Resolves a path to a model-related file, treating \code{x} as a model name. By default (\code{ext} is NULL) the run directory is returned.
}
