% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{modelpath.numeric}
\alias{modelpath.numeric}
\title{Resolve A Path to a Model-related File for Numeric}
\usage{
\method{modelpath}{numeric}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
character
}
\description{
Resolves a path to a model-related file, coercing \code{x} to character.
}
