% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{modelpath}
\alias{modelpath}
\title{Resolve A Path to a Model-related File}
\usage{
modelpath(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
character
}
\description{
Resolves a path to a model-related file.
}
