% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superset.R
\name{ninput.character}
\alias{ninput.character}
\title{Calculate Number of Inputs for Character}
\usage{
\method{ninput}{character}(x, ...)
}
\arguments{
\item{x}{character}

\item{...}{passed arguments}
}
\value{
integer
}
\description{
Calculates number of inputs for character by treating as a model name.
}
