% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superset.R
\name{ninput.numeric}
\alias{ninput.numeric}
\title{Calculate Number of Inputs for Numeric}
\usage{
\method{ninput}{numeric}(x, ...)
}
\arguments{
\item{x}{numeric}

\item{...}{passed arguments}
}
\description{
Calculates number of inputs for numeric by coercing to character.
}
