% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical.R
\name{nms_canonical.character}
\alias{nms_canonical.character}
\title{Generate Canonical Names for Character}
\usage{
\method{nms_canonical}{character}(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed arguments}
}
\description{
Generates canonical names for character by converting to parsed model.
}
\examples{
library(magrittr)
options(project = system.file('project/model',package='nonmemica'))
1001 \%>\% nms_canonical
}
