% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical.R
\name{nms_canonical}
\alias{nms_canonical}
\title{Generate Canonical Names}
\usage{
nms_canonical(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed arguments}
}
\description{
Generates canonical names.
}
\keyword{internal}
