% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical.R
\name{nms_nonmem.nms_canonical}
\alias{nms_nonmem.nms_canonical}
\title{Convert Canonical Parameter Names to NONMEM}
\usage{
\method{nms_nonmem}{nms_canonical}(x, ...)
}
\arguments{
\item{x}{nms_canonical}

\item{...}{ignored}
}
\value{
nms_nonmem
}
\description{
Converts nms_canonical to nms_nonmem.
}
\examples{
# example code

}
\seealso{
Other nms: 
\code{\link{as_nms_canonical}()},
\code{\link{as_nms_nonmem}()},
\code{\link{as_nms_pmx}()},
\code{\link{as_nms_psn}()},
\code{\link{nms_canonical.character}()},
\code{\link{nms_canonical.model}()},
\code{\link{nms_canonical.nms_nonmem}()},
\code{\link{nms_canonical.nms_pmx}()},
\code{\link{nms_canonical.numeric}()},
\code{\link{nms_canonical}()},
\code{\link{nms_nonmem.character}()},
\code{\link{nms_nonmem.model}()},
\code{\link{nms_nonmem.nms_pmx}()},
\code{\link{nms_nonmem.numeric}()},
\code{\link{nms_nonmem}()},
\code{\link{nms_pmx.nms_canonical}()},
\code{\link{nms_pmx.nms_nonmem}()},
\code{\link{nms_pmx}()},
\code{\link{nms_psn.character}()},
\code{\link{nms_psn.model}()},
\code{\link{nms_psn.numeric}()},
\code{\link{nms_psn}()}
}
\concept{nms}
