% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical.R
\name{nms_nonmem}
\alias{nms_nonmem}
\title{Generate NONMEM-style Names}
\usage{
nms_nonmem(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed arguments}
}
\description{
Generates NONMEM-style names.
}
\keyword{internal}
