% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical.R
\name{nms_psn.numeric}
\alias{nms_psn.numeric}
\title{Generate PsN-style Names for Numeric}
\usage{
\method{nms_psn}{numeric}(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed arguments}
}
\description{
Generates PsN-style names for numeric by coercing to character.
}
