% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical.R
\name{num_parameters.default}
\alias{num_parameters.default}
\title{Get the Number of Declared Parameters for a NONMEM Model}
\usage{
\method{num_parameters}{default}(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed arguments}
}
\value{
integer
}
\description{
Gets the number of declared parameters for a NONMEM model.
}
\details{
Parameters fixed to some value are included, but unlike (*.ext) matrix off-diagonals
are not included unless specified.
}
\examples{
library(magrittr)
options(project = system.file('project/model',package='nonmemica'))
1001 \%>\% num_parameters
}
\keyword{internal}
