% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_parameters.R
\name{num_parameters}
\alias{num_parameters}
\title{Get the Number of Parameters}
\usage{
num_parameters(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed arguments}
}
\description{
Gets the Number of Parameters associated with an object.
Generic, with method \code{\link{num_parameters.default}}.
}
\keyword{internal}
