% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halfmatrix.R
\name{offdiag.halfmatrix}
\alias{offdiag.halfmatrix}
\title{Isolate Off-diagonal of Half Matrix}
\usage{
\method{offdiag}{halfmatrix}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Isolates off-diagonal of halfmatrix.
}
\seealso{
Other halfmatrix: 
\code{\link{as.data.frame.halfmatrix}()},
\code{\link{as.halfmatrix.default}()},
\code{\link{as.halfmatrix.halfmatrix}()},
\code{\link{as.halfmatrix}()},
\code{\link{as.matrix.halfmatrix}()},
\code{\link{half.matrix}()},
\code{\link{half}()},
\code{\link{is.square.matrix}()},
\code{\link{is.square}()},
\code{\link{offdiag}()},
\code{\link{ord.halfmatrix}()},
\code{\link{ord.matrix}()},
\code{\link{ord}()},
\code{\link{print.halfmatrix}()}
}
\concept{halfmatrix}
