% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.nmctl.R
\name{ord.initList}
\alias{ord.initList}
\title{Identify the order of an initList}
\usage{
\method{ord}{initList}(x, ...)
}
\arguments{
\item{x}{initList}

\item{...}{dots}
}
\value{
numeric
}
\description{
Identifies the order of an initList.
}
\details{
Essentially the length of the list, or the length of the diagonal of a matrix (if BLOCK was defined).
}
\keyword{internal}

