% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{ord.inits}
\alias{ord.inits}
\title{Identify the order of an inits}
\usage{
\method{ord}{inits}(x, ...)
}
\arguments{
\item{x}{inits}

\item{...}{dots}
}
\value{
numeric
}
\description{
Identifies the order of an inits.
}
\details{
Essentially the length of the list, or the length of the diagonal of a matrix (if BLOCK was defined).
}
\seealso{
Other ord: 
\code{\link{ord.items}()}
}
\concept{ord}
\keyword{internal}
