% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.nmctl.R
\name{ord.itemList}
\alias{ord.itemList}
\title{Identify the order of an itemList}
\usage{
\method{ord}{itemList}(x, ...)
}
\arguments{
\item{x}{itemList}

\item{...}{dots}
}
\value{
numeric
}
\description{
Identifies the order of an itemList.
}
\details{
Essentially the length of the list
}
\keyword{internal}

