% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{ord.items}
\alias{ord.items}
\title{Identify the Order of an Items Object}
\usage{
\method{ord}{items}(x, ...)
}
\arguments{
\item{x}{items}

\item{...}{dots}
}
\value{
numeric
}
\description{
Identifies the order of an items object.
}
\details{
Essentially the length of the list
}
\seealso{
Other ord: \code{\link{ord.inits}}
}
\keyword{internal}
