% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likebut.R
\name{parameters.character}
\alias{parameters.character}
\title{Get Parameters for Character.}
\usage{
\method{parameters}{character}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
data.frame
}
\description{
Gets parameters, treating character as model names. If x is length one, 
slightly more details are returned such as datafile, reference model, and feature.
Otherwise results are bound together, one model per column.
}

