% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likebut.R
\name{parameters.numeric}
\alias{parameters.numeric}
\title{Get Parameters for Numeric}
\usage{
\method{parameters}{numeric}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Gets parameters for numeric by coercing to character.
}
\seealso{
Other parameters: \code{\link{parameters.character}},
  \code{\link{parameters}}
}
\keyword{internal}
