% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likebut.R
\name{parameters}
\alias{parameters}
\title{Get Parameters}
\usage{
parameters(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Gets parameters.
}
\seealso{
Other parameters: \code{\link{parameters.character}},
  \code{\link{parameters.numeric}}
}
\keyword{internal}
