% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{parens}
\alias{parens}
\title{Enclose in Parentheses}
\usage{
parens(x, ...)
}
\arguments{
\item{x}{vector}

\item{...}{dots}
}
\value{
character
}
\description{
Enclose in parentheses
}
\seealso{
Other util: 
\code{\link{\%contains\%}()},
\code{\link{as.best}()},
\code{\link{enclose}()},
\code{\link{locf}()},
\code{\link{padded}()},
\code{\link{pool}()},
\code{\link{runhead}()},
\code{\link{text2decimal}()}
}
\concept{util}
\keyword{internal}
