% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partab.R
\name{partab.numeric}
\alias{partab.numeric}
\title{Create Model Parameter Table from Number.}
\usage{
\method{partab}{numeric}(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{arguments to methods}
}
\description{
Creates a model parameter table from a number.
}
\details{
Just coerces to character and calls partab again.
}
\seealso{
Other partab: 
\code{\link{partab.character}()},
\code{\link{partab}()}
}
\concept{partab}
\keyword{internal}
