% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likebut.R
\name{plot.character}
\alias{plot.character}
\title{Plot Character in Project Context}
\usage{
\method{plot}{character}(x, y, ...)
}
\arguments{
\item{x}{object}

\item{y}{passed to plot}

\item{...}{passed arguments}
}
\description{
Plots in project context by treating character as modelname.  A dataset
is constructed by combining the meta version of the model input with a
meta version of the model output and calling plot with the result.
}
\seealso{
as.pharmval
}

