% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likebut.R
\name{plot.numeric}
\alias{plot.numeric}
\title{Plot Numeric in Project Context}
\usage{
\method{plot}{numeric}(x, y, ...)
}
\arguments{
\item{x}{object}

\item{y}{passed to plot}

\item{...}{passed arguments}
}
\description{
Plots numeric in project context by coercing first to character.
}

