% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{pool}
\alias{pool}
\title{Compare Sets}
\usage{
pool(x, y)
}
\arguments{
\item{x}{vector}

\item{y}{vector}
}
\value{
list: unique x, unique y, and intersection
}
\description{
Compare sets by evaluating both set differences, and the intersection.
}
\examples{
pool(1:3,2:4)
}
\seealso{
Other util: 
\code{\link{\%contains\%}()},
\code{\link{as.best}()},
\code{\link{enclose}()},
\code{\link{locf}()},
\code{\link{padded}()},
\code{\link{parens}()},
\code{\link{runhead}()},
\code{\link{text2decimal}()}
}
\concept{util}
\keyword{internal}
