% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{prettycom.character}
\alias{prettycom.character}
\title{Pretty-print a Comment for Character}
\usage{
\method{prettycom}{character}(x, widths, sep, delim, ...)
}
\arguments{
\item{x}{... object}

\item{widths}{integer}

\item{sep}{input separator}

\item{delim}{output separator}

\item{...}{passed arguments}
}
\description{
Pretty-prints a comment for character.
}
\seealso{
Other prettycom: 
\code{\link{prettycom}()}
}
\concept{prettycom}
\keyword{internal}
