% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{prettycom}
\alias{prettycom}
\title{Pretty-print a Comment}
\usage{
prettycom(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Pretty-prints a comment.
}
\seealso{
Other prettycom: 
\code{\link{prettycom.character}()}
}
\concept{prettycom}
\keyword{internal}
