% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halfmatrix.R
\name{print.halfmatrix}
\alias{print.halfmatrix}
\title{Print Half Matrix}
\usage{
\method{print}{halfmatrix}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Prints half matrix.
}
\seealso{
Other halfmatrix: 
\code{\link{as.data.frame.halfmatrix}()},
\code{\link{as.halfmatrix}()},
\code{\link{as.halfmatrix.default}()},
\code{\link{as.halfmatrix.halfmatrix}()},
\code{\link{as.matrix.halfmatrix}()},
\code{\link{half}()},
\code{\link{half.matrix}()},
\code{\link{is.square}()},
\code{\link{is.square.matrix}()},
\code{\link{offdiag}()},
\code{\link{offdiag.halfmatrix}()},
\code{\link{ord}()},
\code{\link{ord.halfmatrix}()},
\code{\link{ord.matrix}()}
}
\concept{halfmatrix}
\keyword{internal}
