% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{print.init}
\alias{print.init}
\title{Print init}
\usage{
\method{print}{init}(x, ...)
}
\arguments{
\item{x}{init}

\item{...}{dots}
}
\value{
character
}
\description{
Prints init.
}
\seealso{
Other print: 
\code{\link{print.inits}()},
\code{\link{print.items}()},
\code{\link{print.model}()}
}
\concept{print}
\keyword{internal}
