% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.init.R
\name{print.initList}
\alias{print.initList}
\title{Print initList}
\usage{
\method{print}{initList}(x, ...)
}
\arguments{
\item{x}{initList}

\item{...}{dots}
}
\value{
character
}
\description{
Prints initList.
}
\keyword{internal}

