% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{print.inits}
\alias{print.inits}
\title{Print inits}
\usage{
\method{print}{inits}(x, ...)
}
\arguments{
\item{x}{inits}

\item{...}{dots}
}
\value{
character
}
\description{
Prints inits.
}
\seealso{
Other print: \code{\link{print.init}},
  \code{\link{print.items}}, \code{\link{print.model}}
}
\keyword{internal}
