% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.nmctl.R
\name{print.itemList}
\alias{print.itemList}
\title{Print itemList}
\usage{
\method{print}{itemList}(x, ...)
}
\arguments{
\item{x}{itemList}

\item{...}{dots}
}
\value{
character
}
\description{
Prints itemList.
}
\keyword{internal}
