% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{print.items}
\alias{print.items}
\title{Print Items}
\usage{
\method{print}{items}(x, ...)
}
\arguments{
\item{x}{items}

\item{...}{dots}
}
\value{
character
}
\description{
Prints items.
}
\keyword{internal}
