% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{print.model}
\alias{print.model}
\title{Print model}
\usage{
\method{print}{model}(x, ...)
}
\arguments{
\item{x}{model}

\item{...}{dots}
}
\value{
character
}
\description{
Print model.
}
\details{
Formats and prints.
}
\keyword{internal}
