% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.nmctl.R
\name{print.nmctl}
\alias{print.nmctl}
\title{Print nmctl}
\usage{
\method{print}{nmctl}(x, ...)
}
\arguments{
\item{x}{nmctl}

\item{...}{dots}
}
\value{
character
}
\description{
Print nmctl.
}
\details{
Formats and prints.
}
\keyword{internal}

