% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likebut.R
\name{problem.numeric}
\alias{problem.numeric}
\title{Identify the Model Problem Statement for Numeric}
\usage{
\method{problem}{numeric}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Identifies the model problem statement for numeric. Coerces to character.
}
\seealso{
Other problem: 
\code{\link{as.problem}()},
\code{\link{problem.character}()},
\code{\link{problem_}()},
\code{\link{problem}()}
}
\concept{problem}
\keyword{internal}
