% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likebut.R
\name{problem}
\alias{problem}
\title{Identify the Model Problem Statement}
\usage{
problem(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Identifies the model problem statement.
}
\seealso{
Other problem: \code{\link{problem.character}},
  \code{\link{problem.numeric}}, \code{\link{problem_}}
}
\keyword{internal}
