% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likebut.R
\name{problem_}
\alias{problem_}
\title{Identify the Single Model Problem Statement}
\usage{
problem_(x, ...)
}
\arguments{
\item{x}{character}

\item{...}{passed arguments}
}
\description{
Identifies a single model problem statement.
}
\seealso{
Other problem: 
\code{\link{as.problem}()},
\code{\link{problem.character}()},
\code{\link{problem.numeric}()},
\code{\link{problem}()}
}
\concept{problem}
\keyword{internal}
