% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{psn_nested}
\alias{psn_nested}
\title{PsN Model File is Nested
Check whether a particular file extension corresponds to
a file that is nested within a subdirectory
using default PsN conventions.}
\usage{
psn_nested(x, ...)
}
\arguments{
\item{x}{character, a file extension, without dot.}

\item{...}{ignored}
}
\value{
logical
}
\description{
PsN Model File is Nested
Check whether a particular file extension corresponds to
a file that is nested within a subdirectory
using default PsN conventions.
}
\examples{
psn_nested('mod')
}
