% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{read.model}
\alias{read.model}
\title{Read model}
\usage{
read.model(con, parse = FALSE, ...)
}
\arguments{
\item{con}{model connection}

\item{parse}{whether to convert thetas to initList objects}

\item{...}{dots}
}
\value{
character
}
\description{
Read model.
}
\details{
Reads model from a connection.
}
\keyword{internal}
