% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.nmctl.R
\name{read.nmctl}
\alias{read.nmctl}
\title{Read nmctl}
\usage{
read.nmctl(con, parse = FALSE, ...)
}
\arguments{
\item{con}{nmctl connection}

\item{parse}{whether to convert thetas to initList objects}

\item{...}{dots}
}
\value{
character
}
\description{
Read nmctl.
}
\details{
Reads nmctl from a connection.
}
\keyword{internal}

