% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpath.R
\name{readxml}
\alias{readxml}
\title{Read XML with Conditional Filtering}
\usage{
readxml(x, strip.namespace = TRUE, estimation = integer(0), ...)
}
\arguments{
\item{x}{file path}

\item{strip.namespace}{whether to strip e.g. nm: from xml elements}

\item{estimation}{integer indicating which estimation step to select, if multiple. Defaults to last.}

\item{...}{passed arguments}
}
\value{
xml_document
}
\description{
Reads XML with conditional filtering, i.e. selecting among multiple estimation steps.
}
\seealso{
Other xpath: 
\code{\link{as.xml_document}()},
\code{\link{as.xml_document.character}()},
\code{\link{as.xml_document.numeric}()},
\code{\link{filter_estimation}()},
\code{\link{xpath}()},
\code{\link{xpath.default}()},
\code{\link{xpath.xml_document}()}
}
\concept{xpath}
