% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{resolve}
\alias{resolve}
\title{Resolve File Path}
\usage{
resolve(file, dir)
}
\arguments{
\item{file}{path to a file}

\item{dir}{reference directory for a relative file path}
}
\value{
character
}
\description{
Resolves a file path.  Returns the path if absolute.  If
relative, concatenates the directory and file.
}
