% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{runhead}
\alias{runhead}
\title{Find Positions in a Vector that Differ from Previous}
\usage{
runhead(x)
}
\arguments{
\item{x}{vector}
}
\value{
logical
}
\description{
Finds postions in a vector that differ from previous.  First position is defined TRUE.
}
