% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likebut.R
\name{runlog.character}
\alias{runlog.character}
\title{Create a Runlog for Character}
\usage{
\method{runlog}{character}(x, dependencies = F, opt = getOption("project"),
  project = if (is.null(opt)) getwd() else opt, ...)
}
\arguments{
\item{x}{object}

\item{dependencies}{whether to log runs in lineage(s) as well}

\item{opt}{alternative specification of project directory}

\item{project}{direct specification of project directory}

\item{...}{passed arguments}
}
\value{
data.frame
}
\description{
Creates a Runlog for character by treating x as modelname(s).
}

