% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safejoin.R
\name{safe_join}
\alias{safe_join}
\title{Join Data Safely}
\usage{
safe_join(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{arguments to methods}
}
\description{
Joins data safely.  Generic, with method for data.frame.
}
\seealso{
\code{\link{safe_join.data.frame}}

Other safe_join: 
\code{\link{safe_join.data.frame}()}
}
\concept{safe_join}
