% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superset.R
\name{shuffle}
\alias{shuffle}
\title{Move the Columns of a Data Frame Relative to Each Other}
\usage{
shuffle(x, who, after = NA, ...)
}
\arguments{
\item{x}{data.frame}

\item{who}{a character vector of column names to move, or a logical vector of length names(x), or a vector of indices}

\item{after}{column after which to put who: may be character, integer, NA, or NULL}

\item{...}{ignored}
}
\value{
data.frame
}
\description{
Move the columns of a data.frame relative to each other.
}
