% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{specfile.character}
\alias{specfile.character}
\title{Identify the Data Specification File for a Model}
\usage{
\method{specfile}{character}(x, find = "\\\\.csv$", use = ".spec", ...)
}
\arguments{
\item{x}{the model name}

\item{find}{file extension to replace}

\item{use}{file extension to use}

\item{...}{pass ext over-ride default file extension in datafile()}
}
\value{
character
}
\description{
Identifies the data specification file associated with the datafile used by a model. Locates the datafile specified in the control stream, and substitutes a different extension.
}
\seealso{
datafile
}
