% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{specfile.numeric}
\alias{specfile.numeric}
\title{Identify Specfile from Numeric}
\usage{
\method{specfile}{numeric}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{dots}
}
\description{
Identifies specfile from numeric.
}
\keyword{internal}
