% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{specfile}
\alias{specfile}
\title{Identify the Data Specification File for a Model}
\usage{
specfile(x, find = "\\\\.csv$", use = ".spec", ...)
}
\arguments{
\item{x}{the model name}

\item{find}{file extension to replace}

\item{use}{file extension to use}

\item{...}{pass ext over-ride default file extension in datafile()}
}
\value{
character
}
\description{
Identifies the data specification file associated with the datafile used by a model. Identifies the datafile specified in the control stream, and substitutes a different extension.
}
\seealso{
datafile
}
