% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.init.R
\name{[.initList}
\alias{[.initList}
\title{Subset initList}
\usage{
\method{[}{initList}(x, ..., drop = TRUE)
}
\arguments{
\item{x}{initList}

\item{...}{dots}

\item{drop}{logical}
}
\value{
initList
}
\description{
Subsets initList.
}
\keyword{internal}

