% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{[.inits}
\alias{[.inits}
\title{Subset inits}
\usage{
\method{[}{inits}(x, ..., drop = TRUE)
}
\arguments{
\item{x}{inits}

\item{...}{passed arguments}

\item{drop}{logical}
}
\value{
inits
}
\description{
Subsets inits.
}
\concept{inits}
\keyword{internal}
