% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{[.model}
\alias{[.model}
\title{Subset model}
\usage{
\method{[}{model}(x, ..., drop = TRUE)
}
\arguments{
\item{x}{model}

\item{...}{dots}

\item{drop}{passed to subset}
}
\value{
model
}
\description{
Subsets model.
}
\seealso{
Other as.model: 
\code{\link{[[.model}()},
\code{\link{as.model.numeric}()},
\code{\link{as.model}()},
\code{\link{read.model}()},
\code{\link{write.model}()}
}
\concept{as.model}
\keyword{internal}
