% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.nmctl.R
\name{[.nmctl}
\alias{[.nmctl}
\title{Subset nmctl}
\usage{
\method{[}{nmctl}(x, ..., drop = TRUE)
}
\arguments{
\item{x}{nmctl}

\item{...}{dots}

\item{drop}{passed to subset}
}
\value{
nmctl
}
\description{
Subsets nmctl.
}
\keyword{internal}

