% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.nmctl.R
\name{[[.nmctl}
\alias{[[.nmctl}
\title{Select nmctl Element}
\usage{
\method{[[}{nmctl}(x, ..., drop = TRUE)
}
\arguments{
\item{x}{nmctl}

\item{...}{dots}

\item{drop}{passed to element select}
}
\value{
element
}
\description{
Selects nmctl element.
}
\keyword{internal}

