% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superset.R
\name{superset}
\alias{superset}
\title{Coerce to Superset}
\usage{
superset(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Coerces to superset.
}
\details{
Generic, with methods for numeric and character.
}
\seealso{
\code{\link{superset.numeric}} \code{\link{superset.character}}

Other superset: 
\code{\link{generalize}()},
\code{\link{meta.character}()},
\code{\link{meta.numeric}()},
\code{\link{metaplot.character}()},
\code{\link{metaplot.numeric}()},
\code{\link{metaplot_character}()},
\code{\link{metasuperset}()},
\code{\link{meta}()},
\code{\link{ninput.character}()},
\code{\link{ninput.numeric}()},
\code{\link{ninput}()},
\code{\link{shuffle}()},
\code{\link{superset.character}()},
\code{\link{superset.numeric}()},
\code{\link{superspec.character}()},
\code{\link{superspec.numeric}()},
\code{\link{superspec}()}
}
\concept{superset}
\keyword{internal}
