% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superset.R
\name{superspec.numeric}
\alias{superspec.numeric}
\title{Create Specification for Model Inputs and Outputs From Numeric}
\usage{
\method{superspec}{numeric}(x, ...)
}
\arguments{
\item{x}{numeric}

\item{...}{passed arguments}
}
\description{
Create a specification for the result of superset() from numeric by coercing to character.
}
