% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superset.R
\name{superspec}
\alias{superspec}
\title{Create Specification for Model Inputs and Outputs}
\usage{
superspec(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Create a specification for the result of superset().
}
