% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tad.R
\docType{data}
\name{tad1}
\alias{tad1}
\title{A NONMEM-like Dataset}
\usage{
data(tad1)
}
\description{
A dataset showing dose and and observation records for several subjects.
Doses are duplicated across compartments 1 and 2 as for mixed absorption modeling.
}
\details{
\itemize{
  \item C. An exclusion flag, NA by default, or 'C'.
  \item ID. Integer subject identifier.
  \item TIME. Numeric event time (h).
  \item EVID. Event type identifier: observation (0) or dose (1).
  \item CMT. Event compartment: dose (1), central (2) or peripheral (4).
  \item AMT. Amount of dose (mg).
  \item RATE. NONMEM RATE item.
  \item ADDL. Number of additional doses, or NA for observations.
  \item II. Interdose interval for additional doses, or NA for observations.
  \item DV. Observation placeholder.
}
}
