% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{text2decimal}
\alias{text2decimal}
\title{Convert Text to Decimal}
\usage{
text2decimal(x)
}
\arguments{
\item{x}{character}
}
\value{
numeric
}
\description{
Scavenge text for first reasonably-inferred numeric value.
}
