% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tad.R
\name{tod}
\alias{tod}
\title{Calculate Time of Most Recent Dose}
\usage{
tod(x, ref, addl, ii, pre = T, ...)
}
\arguments{
\item{x}{a numeric vector of event times}

\item{ref}{length x vector of reference dose times}

\item{addl}{length x integer: number of additional doses}

\item{ii}{length x numeric: interdose interval for addl}

\item{pre}{assume that simultaneous sample precedes implied dose}

\item{...}{ignored}
}
\value{
numeric
}
\description{
Calculates time of most recent dose.
}
\seealso{
\code{\link{tad}}

Other tad: 
\code{\link{tad}()}
}
\concept{tad}
