% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likebut.R
\name{tweak.default}
\alias{tweak.default}
\title{Tweak a Model}
\usage{
tweak.default(x, opt = getOption("project"), project = if (is.null(opt))
  getwd() else opt, start = NULL, n = 10, extension = ".ctl",
  include = ".def", ...)
}
\arguments{
\item{x}{object}

\item{opt}{alternative specification of project directory}

\item{project}{direct specification of project directory}

\item{start}{a number to use as the first modelname}

\item{n}{the number of variants to generate (named start:n)}

\item{extension}{extension for the model file}

\item{include}{regular expressions for files to copy to new directory}

\item{...}{dots}
}
\value{
character: vector of names for models created
}
\description{
Tweaks a model by jittering initial estimates. Creates a new model directory
in project context and places the model there. Copies helper files as well.
Expects that x is a number.
}

