% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{tweak.inits}
\alias{tweak.inits}
\title{Tweak inits}
\usage{
\method{tweak}{inits}(x, sd = 0.13, digits = 3, ...)
}
\arguments{
\item{x}{object}

\item{sd}{numeric}

\item{digits}{integer}

\item{...}{dots}
}
\value{
inits
}
\description{
Tweaks inits.
}
\seealso{
Other tweak: 
\code{\link{tweak.default}()},
\code{\link{tweak.init}()},
\code{\link{tweak.model}()},
\code{\link{tweak}()}
}
\concept{tweak}
\keyword{internal}
