% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{tweak.model}
\alias{tweak.model}
\title{Tweak model}
\usage{
\method{tweak}{model}(x, sd = 0.13, digits = 3, ...)
}
\arguments{
\item{x}{object}

\item{sd}{numeric}

\item{digits}{integer}

\item{...}{dots}
}
\value{
model
}
\description{
Tweaks model.
}
\seealso{
Other tweak: \code{\link{tweak.inits}},
  \code{\link{tweak.init}}, \code{\link{tweak}}
}
