% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.init.R
\name{tweak.nmctl}
\alias{tweak.nmctl}
\title{Tweak nmctl}
\usage{
\method{tweak}{nmctl}(x, sd = 0.13, digits = 3, ...)
}
\arguments{
\item{x}{object}

\item{sd}{numeric}

\item{digits}{integer}

\item{...}{dots}
}
\value{
nmctl
}
\description{
Tweaks nmctl.
}
\seealso{
Other tweak: \code{\link{tweak.initList}},
  \code{\link{tweak.init}}, \code{\link{tweak}}
}

