% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{tweak}
\alias{tweak}
\title{Tweak Something}
\usage{
tweak(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{dots}
}
\description{
Tweaks something.
}
\seealso{
Other tweak: \code{\link{tweak.inits}},
  \code{\link{tweak.init}}, \code{\link{tweak.model}}
}
