% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superset.R
\name{unfold.character}
\alias{unfold.character}
\title{Unfold Character}
\usage{
\method{unfold}{character}(x, ...)
}
\arguments{
\item{x}{character}

\item{...}{passed arguments}
}
\description{
Unfold by treating character as model name.  A dataset
is constructed by combining the meta version of the model input with a
meta version of the model output and calling unfold with the result.
}
\seealso{
code{fold}
}
