% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superset.R
\name{unfold.numeric}
\alias{unfold.numeric}
\title{Unfold Numeric}
\usage{
\method{unfold}{numeric}(x, ...)
}
\arguments{
\item{x}{numeric}

\item{...}{passed arguments}
}
\description{
Unfolds numeric by coercing to character.
}
\keyword{internal}
