% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{updated.character}
\alias{updated.character}
\title{Create the Updated Version of Character}
\usage{
\method{updated}{character}(x, initial = estimates(x, ...), parse = TRUE, verbose = FALSE, ...)
}
\arguments{
\item{x}{character}

\item{initial}{values to use for initial estimates (numeric)}

\item{parse}{whether to parse the initial estimates, etc.}

\item{verbose}{extended messaging}

\item{...}{dots}
}
\value{
model
}
\description{
Creates the updated version of character by treating as a modelname. Parses the associated control stream and ammends the initial estimates to reflect model results (as per xml file).
}
\seealso{
Other updated: 
\code{\link{updated.numeric}()},
\code{\link{updated}()}
}
\concept{updated}
