% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{updated.numeric}
\alias{updated.numeric}
\title{Create the Updated Version of Numeric}
\usage{
\method{updated}{numeric}(x, ...)
}
\arguments{
\item{x}{numeric}

\item{...}{passed arguments}
}
\description{
Creates the updated version of numeric by coercing to character.
}
\seealso{
Other updated: 
\code{\link{updated}()},
\code{\link{updated.character}()}
}
\concept{updated}
\keyword{internal}
