% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{updated}
\alias{updated}
\title{Create the Updated Version of Something}
\usage{
updated(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{dots}
}
\description{
Creates the updated version of something. Don't confuse with stats::update.
}
\seealso{
Other updated: \code{\link{updated.character}},
  \code{\link{updated.numeric}}
}
\keyword{internal}
