% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getset.R
\name{upper<-.model}
\alias{upper<-.model}
\title{Set Upper Bounds for Model Initial Estimates}
\usage{
\method{upper}{model}(x) <- value
}
\arguments{
\item{x}{model}

\item{value}{numeric}
}
\description{
Sets upper bounds for model initial estimates.
}
\seealso{
Other upper: \code{\link{upper.model}},
  \code{\link{upper<-}}, \code{\link{upper}}
}
