% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getset.R
\name{upper<-.nmctl}
\alias{upper<-.nmctl}
\title{Set Upper Bounds for Model Initial Estimates}
\usage{
\method{upper}{nmctl}(x) <- value
}
\arguments{
\item{x}{nmctl}

\item{value}{numeric}
}
\description{
Sets upper bounds for model initial estimates.
}

