% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getset.R
\name{upper.model}
\alias{upper.model}
\title{Get Upper Bounds for Model Initial Estimates}
\usage{
\method{upper}{model}(x, ...)
}
\arguments{
\item{x}{model}

\item{...}{dots}
}
\description{
Gets upper bounds for model initial estimates.
}
\examples{
library(magrittr)
options(project = system.file('project/model',package='nonmemica'))
1001 \%>\% as.model \%>\% upper
}
\seealso{
Other upper: 
\code{\link{upper<-.model}()},
\code{\link{upper<-}()},
\code{\link{upper}()}
}
\concept{upper}
