% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getset.R
\name{upper.nmctl}
\alias{upper.nmctl}
\title{Get Upper Bounds for Model Initial Estimates}
\usage{
\method{upper}{nmctl}(x, ...)
}
\arguments{
\item{x}{nmctl}

\item{...}{dots}
}
\description{
Gets upper bounds for model initial estimates.
}

