% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getset.R
\name{upper}
\alias{upper}
\title{Get Upper Value}
\usage{
upper(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{dots}
}
\description{
Gets upper Value.
}
\seealso{
Other upper: 
\code{\link{upper.model}()},
\code{\link{upper<-.model}()},
\code{\link{upper<-}()}
}
\concept{upper}
\keyword{internal}
