% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partab.R
\name{val_name}
\alias{val_name}
\title{Create Model Parameter Table from xml_document
Creates a model parameter table from xml_document, evaluating the xpath expression and supplying arbitrary names for parameter and moment of parameter.}
\usage{
val_name(x, xpath, param, moment, ...)
}
\arguments{
\item{x}{xml_document}

\item{xpath}{character}

\item{param}{character}

\item{moment}{character}

\item{...}{passed arguments}
}
\value{
data.frame
}
\description{
Create Model Parameter Table from xml_document
Creates a model parameter table from xml_document, evaluating the xpath expression and supplying arbitrary names for parameter and moment of parameter.
}
