% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{write.model}
\alias{write.model}
\title{Write model}
\usage{
write.model(x, file = "data", ncolumns = 1, append = FALSE, sep = " ", ...)
}
\arguments{
\item{x}{model}

\item{file}{passed to write()}

\item{ncolumns}{passed to write()}

\item{append}{passed to write()}

\item{sep}{passed to write()}

\item{...}{dots}
}
\value{
used for side effects
}
\description{
Write model.
}
\details{
writes (formatted) model to file.
}
\seealso{
Other as.model: 
\code{\link{[.model}()},
\code{\link{[[.model}()},
\code{\link{as.model.numeric}()},
\code{\link{as.model}()},
\code{\link{read.model}()}
}
\concept{as.model}
\keyword{internal}
