% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.nmctl.R
\name{write.nmctl}
\alias{write.nmctl}
\title{Write nmctl}
\usage{
write.nmctl(x, file = "data", ncolumns = 1, append = FALSE, sep = " ",
  ...)
}
\arguments{
\item{x}{nmctl}

\item{file}{passed to write()}

\item{ncolumns}{passed to write()}

\item{append}{passed to write()}

\item{sep}{passed to write()}

\item{...}{dots}
}
\value{
used for side effects
}
\description{
Write nmctl.
}
\details{
writes (formatted) nmctl to file.
}
\keyword{internal}

