% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.init.R
\name{as.character.init}
\alias{as.character.init}
\title{Coerce init to character}
\usage{
\method{as.character}{init}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{dots}
}
\value{
character
}
\description{
Coerces init to character.
}
\keyword{internal}

