% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{as.character.items}
\alias{as.character.items}
\title{Coerce Items to Character}
\usage{
\method{as.character}{items}(
  x,
  pretty = TRUE,
  sep = ";",
  delim = " ; ",
  widths = comwidth(x),
  ...
)
}
\arguments{
\item{x}{items}

\item{pretty}{logical}

\item{sep}{input delimiter}

\item{delim}{output delimiter}

\item{widths}{desired widths}

\item{...}{passed arguments}
}
\value{
character
}
\description{
Coerces items to character.
}
\seealso{
Other as.character: 
\code{\link{as.character.inits}()},
\code{\link{as.character.init}()},
\code{\link{as.character.model}()},
\code{\link{as.character.problem}()}
}
\concept{as.character}
\keyword{internal}
