% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{as.init}
\alias{as.init}
\alias{as.init.init}
\title{Coerce to init}
\usage{
as.init(x, ...)

\method{as.init}{init}(x = numeric(0), fixed = FALSE, comment = character(0), ...)
}
\arguments{
\item{x}{object}

\item{...}{dots}

\item{fixed}{logical}

\item{comment}{character}
}
\value{
init
}
\description{
Coerces to init

Coerces init to init
}
\section{Methods (by class)}{
\itemize{
\item \code{init}: init method
}}

\seealso{
Other as.init: 
\code{\link{as.init.numeric}()}

Other as.init: 
\code{\link{as.init.numeric}()}
}
\concept{as.init}
\keyword{internal}
