% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.init.R
\name{as.init}
\alias{as.init}
\alias{as.init.init}
\alias{as.init.numeric}
\title{Coerce to init}
\usage{
as.init(x, ...)

\method{as.init}{init}(x = numeric(0), fixed = FALSE,
  comment = character(0), ...)

\method{as.init}{numeric}(x = numeric(0), fixed = FALSE,
  comment = character(0), ...)
}
\arguments{
\item{x}{object}

\item{...}{dots}

\item{fixed}{logical}

\item{comment}{character}

\item{...}{dots}

\item{fixed}{logical}

\item{comment}{character}

\item{...}{dots}
}
\value{
init

init
}
\description{
Coerces to init

Coerces init to init

Coerces numeric to init
}
\section{Methods (by class)}{
\itemize{
\item \code{init}: init method

\item \code{numeric}: numeric method
}}
\keyword{internal}

