% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{as.inits.numeric}
\alias{as.inits.numeric}
\title{Coerce numeric to inits}
\usage{
\method{as.inits}{numeric}(x, fixed = FALSE, comment = character(0), ...)
}
\arguments{
\item{x}{object}

\item{fixed}{logical}

\item{comment}{character}

\item{...}{dots}
}
\value{
inits
}
\description{
Coerces numeric to inits
}
\keyword{internal}
