% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.nmctl.R
\name{as.itemComments}
\alias{as.itemComments}
\alias{as.itemComments.initList}
\alias{as.itemComments.itemList}
\alias{as.itemComments.nmctl}
\alias{as.itemComments.nmctlType}
\title{Coerce to itemComments}
\usage{
as.itemComments(x, ...)

\method{as.itemComments}{nmctlType}(x, ...)

\method{as.itemComments}{nmctl}(x, fields = c("symbol", "unit", "label"),
  expected = character(0), na = NA_character_, ...)

\method{as.itemComments}{itemList}(x, type, prior, ...)

\method{as.itemComments}{initList}(x, type, prior, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{dots}

\item{fields}{data items to scavenge from control stream comments}

\item{expected}{parameters known from NONMEM output}

\item{na}{string to use for NA values when writing default metafile}

\item{type}{item type: table}

\item{prior}{number of prior items of this type (maybe imporant for numbering)}

\item{type}{item type: theta, omega, sigma, or table}
}
\value{
data.frame

data.frame

data.frame

data.frame
}
\description{
Coerces to itemComments

Converts nmctlType to itemComments

Converts nmctl to itemComments

Converts itemList to itemComments

Converts initList to itemComments
}
\section{Methods (by class)}{
\itemize{
\item \code{nmctlType}: nmctlType method

\item \code{nmctl}: nmctl method

\item \code{itemList}: initList method

\item \code{initList}: initList method
}}
\keyword{internal}

