% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{as.items.character}
\alias{as.items.character}
\title{Convert to Items from Character}
\usage{
\method{as.items}{character}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
items
}
\description{
Converts to items from character
}
