% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpath.R
\name{as.xml_document.numeric}
\alias{as.xml_document.numeric}
\title{Coerce numeric to xml_document}
\usage{
\method{as.xml_document}{numeric}(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{arguments to methods}
}
\value{
xml_document
}
\description{
Coerces numeric to xml_document
}
\seealso{
Other xpath: 
\code{\link{as.xml_document}()},
\code{\link{xpath}()}
}
\concept{xpath}
\keyword{internal}
