% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpath.R
\name{as.xml_document}
\alias{as.xml_document}
\alias{as.xml_document.xml_document}
\title{Create an xml_document in a Project Context}
\usage{
as.xml_document(x, ...)

\method{as.xml_document}{xml_document}(x, ...)
}
\arguments{
\item{x}{xml_document}

\item{...}{ignored}
}
\value{
xml_document

xml_document
}
\description{
Creates an xml_document in a project context.

Coerces xml_document to xml_document
}
\section{Methods (by class)}{
\itemize{
\item \code{as.xml_document(xml_document)}: xml_document method

}}
\examples{
library(magrittr)
options(project = system.file('project/model',package='nonmemica'))
1001 \%>\% as.xml_document
}
\seealso{
\code{\link{xpath}}

Other xpath: 
\code{\link{as.xml_document.character}()},
\code{\link{as.xml_document.numeric}()},
\code{\link{xpath.default}()},
\code{\link{xpath.xml_document}()},
\code{\link{xpath}()}

Other xpath: 
\code{\link{as.xml_document.character}()},
\code{\link{as.xml_document.numeric}()},
\code{\link{xpath.default}()},
\code{\link{xpath.xml_document}()},
\code{\link{xpath}()}
}
\concept{xpath}
