% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpath.R
\name{as.xml_document}
\alias{as.xml_document}
\alias{as.xml_document.xml_document}
\alias{as.xml_document.character}
\title{Create an xml_document in a Project Context}
\usage{
as.xml_document(x, ...)

\method{as.xml_document}{xml_document}(x, ...)

\method{as.xml_document}{character}(x, strip.namespace = TRUE, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{arguments to methods}

\item{strip.namespace}{whether to strip e.g. nm: from xml elements}
}
\value{
xml_document

xml_document

xml_document
}
\description{
Creates an xml_document in a project context.

Coerces xml_document to xml_document

Creates an xml_document from character (modelname or filepath).
}
\section{Methods (by class)}{
\itemize{
\item \code{xml_document}: xml_document method

\item \code{character}: filepath method
}}

\examples{
library(magrittr)
options(project = system.file('project/model',package='nonmemica'))
1001 \%>\% as.xml_document
}
\seealso{
\code{\link{xpath}}

Other xpath: 
\code{\link{as.xml_document.numeric}()},
\code{\link{xpath}()}

Other xpath: 
\code{\link{as.xml_document.numeric}()},
\code{\link{xpath}()}

Other xpath: 
\code{\link{as.xml_document.numeric}()},
\code{\link{xpath}()}
}
\concept{xpath}
