% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likebut.R
\name{but.default}
\alias{but.default}
\title{Identify the Distinctive Feature of a Model}
\usage{
\method{but}{default}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
character
}
\description{
Identifies the distinctive feature of a model, i.e. how it differs from its parent (the reference model). x can be a vector.
}
\seealso{
Other but: 
\code{\link{but}()}
}
\concept{but}
\keyword{internal}
