% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.init.R
\name{$<-.init}
\alias{$<-.init}
\title{Set init element}
\usage{
\method{$}{init}(x, name) <- value
}
\arguments{
\item{x}{init}

\item{name}{character}

\item{value}{numeric}
}
\value{
init
}
\description{
Sets init element.
}
\keyword{internal}

