% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{comwidth.inits}
\alias{comwidth.inits}
\title{Calculate Comment Widths for Inits}
\usage{
\method{comwidth}{inits}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Calculates comment widths for inits
}
\keyword{internal}
