% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{datafile.numeric}
\alias{datafile.numeric}
\title{Identify Datafile from Numeric}
\usage{
\method{datafile}{numeric}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{dots}
}
\description{
Identifies datafile from numeric.
}
\seealso{
Other path: 
\code{\link{datafile.character}()},
\code{\link{datafile}()},
\code{\link{modeldir}()},
\code{\link{modelfile}()},
\code{\link{modelpath.character}()},
\code{\link{modelpath.numeric}()},
\code{\link{modelpath}()},
\code{\link{psn_options}()},
\code{\link{specfile.character}()},
\code{\link{specfile.numeric}()},
\code{\link{specfile}()}
}
\concept{path}
\keyword{internal}
