% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{datafile}
\alias{datafile}
\title{Identify the Datafile for a Model}
\usage{
datafile(x, ...)
}
\arguments{
\item{x}{the model name or path to a control stream}

\item{...}{ext can be passed to modelfile, etc.}
}
\value{
character
}
\description{
Identifies the datafile used by a model.
}
