% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likebut.R
\name{datafile}
\alias{datafile}
\title{Identify the Datafile for a Model}
\usage{
datafile(x, ..., opt = getOption("project"), project = if (is.null(opt))
  getwd() else opt)
}
\arguments{
\item{x}{the model name}

\item{...}{passed arguments}

\item{opt}{alternative specification of project directory}

\item{project}{direct specification of project directory}
}
\value{
character
}
\description{
Identifies the datafile used by a model.
}

