% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likebut.R
\name{depends.default}
\alias{depends.default}
\title{Identify Model Dependencies}
\usage{
\method{depends}{default}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
character
}
\description{
Identify those models in the lineage of models in x.
}
\seealso{
Other depends: \code{\link{depends}}
}
