% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.init.R
\name{fixed<-.init}
\alias{fixed<-.init}
\title{Set fixed init value}
\usage{
\method{fixed}{init}(x) <- value
}
\arguments{
\item{x}{object}

\item{value}{value to}
}
\value{
init
}
\description{
Sets fixed init value.
}
\keyword{internal}

