% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.init.R
\name{format.init}
\alias{format.init}
\title{Format init}
\usage{
\method{format}{init}(x, ...)
}
\arguments{
\item{x}{init}

\item{...}{dots}
}
\value{
character
}
\description{
Formats init.
}
\keyword{internal}

