% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{format.inits}
\alias{format.inits}
\title{Format inits}
\usage{
\method{format}{inits}(x, ...)
}
\arguments{
\item{x}{inits}

\item{...}{dots}
}
\value{
character
}
\description{
Formats inits.
}
\keyword{internal}
