% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{\%contains\%}
\alias{\%contains\%}
\title{Check Whether x contains y}
\usage{
x \%contains\% y
}
\arguments{
\item{x}{character vector to check}

\item{y}{pattern}
}
\value{
logical
}
\description{
Checks whether x contains y, in natural syntax.
}
\examples{
letters \%contains\% 'a'
}
\seealso{
\code{\link{contains}}

Other util: 
\code{\link{as.best}()},
\code{\link{enclose}()},
\code{\link{locf}()},
\code{\link{padded}()},
\code{\link{parens}()},
\code{\link{pool}()},
\code{\link{runhead}()},
\code{\link{text2decimal}()}
}
\concept{util}
\keyword{internal}
