% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getset.R
\name{initial.model}
\alias{initial.model}
\title{Get Model Initial Estimates}
\usage{
\method{initial}{model}(x, ...)
}
\arguments{
\item{x}{model}

\item{...}{dots}
}
\description{
Gets model initial estimates.
}
\examples{
library(magrittr)
options(project = system.file('project/model',package='nonmemica'))
1001 \%>\% as.model \%>\% initial
}
\seealso{
Other initial: 
\code{\link{initial<-.model}()},
\code{\link{initial<-}()},
\code{\link{initial}()}
}
\concept{initial}
