% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{initSubscripts}
\alias{initSubscripts}
\title{Identify Subscripts}
\usage{
initSubscripts(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed arguments}
}
\description{
Identifies subscripts.
}
\seealso{
Other initSubscripts: 
\code{\link{initSubscripts.model}()}
}
\concept{initSubscripts}
\keyword{internal}
