% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.nmctl.R
\name{initSubscripts}
\alias{initSubscripts}
\title{Identify Subscripts}
\usage{
initSubscripts(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{dots}
}
\description{
Identifies subscripts.
}
\keyword{internal}
