% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likebut.R
\name{like.default}
\alias{like.default}
\title{Identify the Relevant Reference Model}
\usage{
\method{like}{default}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
character
}
\description{
Identifies the relevant reference model, i.e. the 'parent' of x.
}
\seealso{
Other like: \code{\link{like}}
}
\keyword{internal}
