% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likebut.R
\name{likebut}
\alias{likebut}
\title{Modify a Model}
\usage{
likebut(x, but = "better", y = NULL, project = getOption("project",
  getwd()), nested = getOption("nested", TRUE), overwrite = FALSE,
  ext = getOption("modex", "ctl"), include = "\\\\.def$", ...)
}
\arguments{
\item{x}{a model name, presumably interpretable as numeric}

\item{but}{a short description of the characteristic difference from x}

\item{y}{optional name for model to be created}

\item{project}{project directory}

\item{nested}{model files nested in run-specific directories}

\item{overwrite}{whether to overwrite y if it exists}

\item{ext}{extension for the model file}

\item{include}{regular expressions for files to copy to new directory}

\item{...}{passed arguments}
}
\value{
the value of y
}
\description{
Makes a copy of a model in a corresponding directory.  Problem statement is
updated to reflect that the model is LIKE the reference model BUT different
in some fundamental way.
}
