% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superset.R
\name{metaplot.character}
\alias{metaplot.character}
\title{Metaplot Character}
\usage{
\method{metaplot}{character}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Plots character by treating as model name.  A dataset
is constructed by combining the meta version of the model input with a
meta version of the model output and calling metaplot with the result.
}
\seealso{
\code{fold}
}
