% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superset.R
\name{metaplot.numeric}
\alias{metaplot.numeric}
\title{Metaplot Numeric}
\usage{
\method{metaplot}{numeric}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Plots numeric by coercing first to character.
}
\seealso{
Other superset: 
\code{\link{generalize}()},
\code{\link{meta.character}()},
\code{\link{meta.numeric}()},
\code{\link{metaplot.character}()},
\code{\link{metaplot_character}()},
\code{\link{metasuperset}()},
\code{\link{meta}()},
\code{\link{ninput.character}()},
\code{\link{ninput.numeric}()},
\code{\link{ninput}()},
\code{\link{shuffle}()},
\code{\link{superset.character}()},
\code{\link{superset.numeric}()},
\code{\link{superset}()},
\code{\link{superspec.character}()},
\code{\link{superspec.numeric}()},
\code{\link{superspec}()}
}
\concept{superset}
\keyword{internal}
