% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superset.R
\name{metaplot_character}
\alias{metaplot_character}
\title{Metaplot Character, Standard Evaluation}
\usage{
metaplot_character(x, groups, meta = NULL, subset, var, ...)
}
\arguments{
\item{x}{object}

\item{groups}{columns by which to group the dataset}

\item{meta}{metadata; meta(x) by default}

\item{subset}{a condition for filtering data}

\item{var}{variables to plot}

\item{...}{passed arguments}
}
\description{
Plots character by treating as model name.  A dataset
is constructed by combining the model input with a
the model output and calling metaplot with the result.
}
\seealso{
Other superset: 
\code{\link{generalize}()},
\code{\link{meta.character}()},
\code{\link{meta.numeric}()},
\code{\link{metaplot.character}()},
\code{\link{metaplot.numeric}()},
\code{\link{metasuperset}()},
\code{\link{meta}()},
\code{\link{ninput.character}()},
\code{\link{ninput.numeric}()},
\code{\link{ninput}()},
\code{\link{shuffle}()},
\code{\link{superset.character}()},
\code{\link{superset.numeric}()},
\code{\link{superset}()},
\code{\link{superspec.character}()},
\code{\link{superspec.numeric}()},
\code{\link{superspec}()}
}
\concept{superset}
