% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superset.R
\name{metasuperset}
\alias{metasuperset}
\title{Retrieve Model Outputs in Meta Format}
\usage{
metasuperset(x, opt = getOption("project"), project = if (is.null(opt))
  getwd() else opt, group_by, meta = match.fun("meta")(x, ...), subset, ...)
}
\arguments{
\item{x}{model name}

\item{opt}{alternative specification of project directory}

\item{project}{direct specification of project directory}

\item{group_by}{vector of key column names in superset, e.g. USUBJID, TIME}

\item{meta}{pre-folded metadata}

\item{subset}{length-one character: a condition for filtering results, e.g. 'EVID == 0'}

\item{...}{passed arguments}
}
\value{
meta
}
\description{
Retrieves model outputs in meta format. Inputs should have EVID.
}
