% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superset.R
\name{metasuperset}
\alias{metasuperset}
\title{Retrieve Model Outputs in Meta Format}
\usage{
metasuperset(x, groups, meta = match.fun("meta")(x, ...), subset, ...)
}
\arguments{
\item{x}{model name}

\item{groups}{vector of key column names in superset, e.g. USUBJID, TIME}

\item{meta}{pre-folded metadata}

\item{subset}{length-one character: a condition for filtering results, e.g. 'EVID == 0'}

\item{...}{passed arguments}
}
\value{
folded
}
\description{
Retrieves model outputs in meta format.
}
