% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{modelfile}
\alias{modelfile}
\title{Identify the Modelfile for a Model}
\usage{
modelfile(x, ext = getOption("modex", "ctl"), ...)
}
\arguments{
\item{x}{the model name}

\item{ext}{model file extension}

\item{...}{passed arguments}
}
\value{
character
}
\description{
Identifies the modelfile used by a model.
}
\examples{
library(magrittr)
options(project = system.file('project/model',package='nonmemica'))
1001 \%>\% modelfile('xml')
}
\seealso{
Other path: 
\code{\link{datafile.character}()},
\code{\link{datafile.numeric}()},
\code{\link{datafile}()},
\code{\link{modeldir}()},
\code{\link{modelpath.character}()},
\code{\link{modelpath.numeric}()},
\code{\link{modelpath}()},
\code{\link{psn_options}()},
\code{\link{specfile.character}()},
\code{\link{specfile.numeric}()},
\code{\link{specfile}()}
}
\concept{path}
