% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{modelpath.numeric}
\alias{modelpath.numeric}
\title{Resolve A Path to a Model-related File for Numeric}
\usage{
\method{modelpath}{numeric}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
character
}
\description{
Resolves a path to a model-related file, coercing \code{x} to character.
}
\seealso{
Other path: 
\code{\link{datafile.character}()},
\code{\link{datafile.numeric}()},
\code{\link{datafile}()},
\code{\link{modeldir}()},
\code{\link{modelfile}()},
\code{\link{modelpath.character}()},
\code{\link{modelpath}()},
\code{\link{psn_options}()},
\code{\link{specfile.character}()},
\code{\link{specfile.numeric}()},
\code{\link{specfile}()}
}
\concept{path}
\keyword{internal}
