% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{modelpath}
\alias{modelpath}
\title{Resolve A Path to a Model-related File}
\usage{
modelpath(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
character
}
\description{
Resolves a path to a model-related file.
}
\examples{
library(magrittr)
options(project = system.file('project/model',package='nonmemica'))
1001 \%>\% modelpath
}
\seealso{
Other path: 
\code{\link{datafile.character}()},
\code{\link{datafile.numeric}()},
\code{\link{datafile}()},
\code{\link{modeldir}()},
\code{\link{modelfile}()},
\code{\link{modelpath.character}()},
\code{\link{modelpath.numeric}()},
\code{\link{psn_options}()},
\code{\link{specfile.character}()},
\code{\link{specfile.numeric}()},
\code{\link{specfile}()}
}
\concept{path}
