% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superset.R
\name{ninput.character}
\alias{ninput.character}
\title{Calculate Number of Inputs for Character}
\usage{
\method{ninput}{character}(x, ...)
}
\arguments{
\item{x}{character}

\item{...}{passed arguments}
}
\value{
integer
}
\description{
Calculates number of inputs for character by treating as a model name.
}
\seealso{
Other superset: 
\code{\link{generalize}()},
\code{\link{meta.character}()},
\code{\link{meta.numeric}()},
\code{\link{metaplot.character}()},
\code{\link{metaplot.numeric}()},
\code{\link{metaplot_character}()},
\code{\link{metasuperset}()},
\code{\link{meta}()},
\code{\link{ninput.numeric}()},
\code{\link{ninput}()},
\code{\link{shuffle}()},
\code{\link{superset.character}()},
\code{\link{superset.numeric}()},
\code{\link{superset}()},
\code{\link{superspec.character}()},
\code{\link{superspec.numeric}()},
\code{\link{superspec}()}
}
\concept{superset}
