% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superset.R
\name{ninput.numeric}
\alias{ninput.numeric}
\title{Calculate Number of Inputs for Numeric}
\usage{
\method{ninput}{numeric}(x, ...)
}
\arguments{
\item{x}{numeric}

\item{...}{passed arguments}
}
\description{
Calculates number of inputs for numeric by coercing to character.
}
\seealso{
Other superset: 
\code{\link{generalize}()},
\code{\link{meta.character}()},
\code{\link{meta.numeric}()},
\code{\link{metaplot.character}()},
\code{\link{metaplot.numeric}()},
\code{\link{metaplot_character}()},
\code{\link{metasuperset}()},
\code{\link{meta}()},
\code{\link{ninput.character}()},
\code{\link{ninput}()},
\code{\link{shuffle}()},
\code{\link{superset.character}()},
\code{\link{superset.numeric}()},
\code{\link{superset}()},
\code{\link{superspec.character}()},
\code{\link{superspec.numeric}()},
\code{\link{superspec}()}
}
\concept{superset}
