% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical.R
\name{nms_canonical.model}
\alias{nms_canonical.model}
\title{Generate Canonical Names for Model}
\usage{
\method{nms_canonical}{model}(x, ...)
}
\arguments{
\item{x}{a model designator}

\item{...}{passed arguments}
}
\value{
canonical (character)
}
\description{
Generates canonical names for a NONMEM control stream object. Canonical names indicate all and only the declared model parameters in lower-case conventional order (theta, omega row-major, sigma) with underscores and two-digit (or more) indices. E.g. theta_01, theta_02, omega_01_01, omega_02_01, omega_02_02, omega_01_01.
}
\seealso{
as.model
}
