% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical.R
\name{nms_nonmem.character}
\alias{nms_nonmem.character}
\title{Generate NONMEM-style Names for Character}
\usage{
\method{nms_nonmem}{character}(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed arguments}
}
\description{
Generates NONMEM-style names for numeric by converting to parsed model.
}
\examples{
library(magrittr)
options(project = system.file('project/model',package='nonmemica'))
1001 \%>\% nms_nonmem
}
\seealso{
Other nms_nonmem: 
\code{\link{nms_nonmem.model}()},
\code{\link{nms_nonmem.numeric}()},
\code{\link{nms_nonmem}()}
}
\concept{nms_nonmem}
