% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical.R
\name{nms_nonmem.model}
\alias{nms_nonmem.model}
\title{Generate NONMEM-style Names for Model}
\usage{
\method{nms_nonmem}{model}(x, ...)
}
\arguments{
\item{x}{a model designator}

\item{...}{passed arguments}
}
\value{
nonmem (character)
}
\description{
Generates NONMEM-style names for parameters declared in a NONMEM control stream object. PsN uses NONMEM-style names, substituting a comment, if any: everything after the first semicolon, up to the second semicolon if present, without leading/trailing spaces/tabs.
}
\seealso{
as.model

Other nms_nonmem: 
\code{\link{nms_nonmem.character}()},
\code{\link{nms_nonmem.numeric}()},
\code{\link{nms_nonmem}()}
}
\concept{nms_nonmem}
