% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical.R
\name{nms_nonmem.numeric}
\alias{nms_nonmem.numeric}
\title{Generate NONMEM-style Names for Numeric}
\usage{
\method{nms_nonmem}{numeric}(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed arguments}
}
\description{
Generates NONMEM-style names for numeric by coercing to character.
}
\seealso{
Other nms_nonmem: 
\code{\link{nms_nonmem.character}()},
\code{\link{nms_nonmem.model}()},
\code{\link{nms_nonmem}()}
}
\concept{nms_nonmem}
\keyword{internal}
