% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical.R
\name{nms_psn.character}
\alias{nms_psn.character}
\title{Generate PsN-style Names for Character}
\usage{
\method{nms_psn}{character}(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed arguments}
}
\description{
Generates PsN-style names for numeric by converting to parsed model.
}
\examples{
library(magrittr)
options(project = system.file('project/model',package='nonmemica'))
1001 \%>\% nms_psn
}
