% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical.R
\name{nms_psn.model}
\alias{nms_psn.model}
\title{Generate PsN-style Names for Model}
\usage{
\method{nms_psn}{model}(x, ...)
}
\arguments{
\item{x}{a model designator}

\item{...}{passed arguments}
}
\value{
psn (character)
}
\description{
Generates PsN-style names for parameters declared in a NONMEM control stream object. PsN uses NONMEM-style names, substituting a comment, if any: everything after the first semicolon, up to the second semicolon if present, without leading/trailing spaces/tabs.
}
\seealso{
as.model
}
