% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halfmatrix.R
\name{ord.halfmatrix}
\alias{ord.halfmatrix}
\title{Calculate Order for Half Matrix}
\usage{
\method{ord}{halfmatrix}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Calculates order for half matrix. I.e., length of one dimension of the equivalent square matrix.
}
\seealso{
Other halfmatrix: 
\code{\link{as.data.frame.halfmatrix}()},
\code{\link{as.halfmatrix.default}()},
\code{\link{as.halfmatrix.halfmatrix}()},
\code{\link{as.halfmatrix}()},
\code{\link{as.matrix.halfmatrix}()},
\code{\link{half.matrix}()},
\code{\link{half}()},
\code{\link{is.square.matrix}()},
\code{\link{is.square}()},
\code{\link{offdiag.halfmatrix}()},
\code{\link{offdiag}()},
\code{\link{ord.matrix}()},
\code{\link{ord}()},
\code{\link{print.halfmatrix}()}
}
\concept{halfmatrix}
\keyword{internal}
