% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{padded}
\alias{padded}
\title{Pad Numeric with Zeros}
\usage{
padded(x, width = 4, ...)
}
\arguments{
\item{x}{numeric}

\item{width}{desired number of characters}

\item{...}{ignored}
}
\value{
character
}
\description{
Pad a numeric vector with leading zeros.
}
\seealso{
Other util: 
\code{\link{\%contains\%}()},
\code{\link{as.best}()},
\code{\link{enclose}()},
\code{\link{locf}()},
\code{\link{parens}()},
\code{\link{pool}()},
\code{\link{runhead}()},
\code{\link{text2decimal}()}
}
\concept{util}
\keyword{internal}
