% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{print.items}
\alias{print.items}
\title{Print Items}
\usage{
\method{print}{items}(x, ...)
}
\arguments{
\item{x}{items}

\item{...}{passed arguments}
}
\value{
character
}
\description{
Prints items.
}
\seealso{
Other print: 
\code{\link{print.inits}()},
\code{\link{print.init}()},
\code{\link{print.model}()}
}
\concept{print}
\keyword{internal}
