% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{print.model}
\alias{print.model}
\title{Print model}
\usage{
\method{print}{model}(x, ...)
}
\arguments{
\item{x}{model}

\item{...}{passed arguments}
}
\value{
character
}
\description{
Print model.
}
\details{
Formats and prints.
}
\seealso{
Other print: 
\code{\link{print.inits}()},
\code{\link{print.init}()},
\code{\link{print.items}()}
}
\concept{print}
\keyword{internal}
