% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likebut.R
\name{problem.character}
\alias{problem.character}
\title{Identify the Model Problem Statement for Character}
\usage{
\method{problem}{character}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
character
}
\description{
Identifies the model problem statement for character (model name).
}
\seealso{
Other problem: 
\code{\link{as.problem}()},
\code{\link{problem.numeric}()},
\code{\link{problem_}()},
\code{\link{problem}()}
}
\concept{problem}
