% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partab.R
\name{row_col}
\alias{row_col}
\title{Create Indexed Model Parameter Table from Xml_document}
\usage{
row_col(x, xpath, param, moment, ...)
}
\arguments{
\item{x}{xml_document}

\item{xpath}{character}

\item{param}{character}

\item{moment}{character}

\item{...}{passed arguments}
}
\value{
data.frame
}
\description{
Creates a model parameter table from xml_document.
}
\details{
The function evaluates the xpath expression, supplying parameter 
and moment labels, and documenting indices of rows and columns.
}
\concept{row_col}
\keyword{internal}
