% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likebut.R
\name{runlog.character}
\alias{runlog.character}
\title{Create a Runlog for Character}
\usage{
\method{runlog}{character}(x, dependencies = F, digits = 3, places = 0,
  ...)
}
\arguments{
\item{x}{object}

\item{dependencies}{whether to log runs in lineage(s) as well}

\item{digits}{significance for parameters}

\item{places}{rounding for objective function}

\item{...}{passed arguments}
}
\value{
data.frame
}
\description{
Creates a Runlog for character by treating x as modelname(s).
}
