% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likebut.R
\name{runlog.numeric}
\alias{runlog.numeric}
\title{Create a Runlog for Numeric}
\usage{
\method{runlog}{numeric}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Creates a runlog for numeric by coercing to character.
}
\keyword{internal}
