% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superset.R
\name{shuffle}
\alias{shuffle}
\title{Move the Columns of a Data Frame Relative to Each Other}
\usage{
shuffle(x, who, after = NA, ...)
}
\arguments{
\item{x}{data.frame}

\item{who}{a character vector of column names to move, or a logical vector of length names(x), or a vector of indices}

\item{after}{column after which to put who: may be character, integer, NA, or NULL}

\item{...}{ignored}
}
\value{
data.frame
}
\description{
Move the columns of a data.frame relative to each other.
}
\seealso{
Other superset: 
\code{\link{generalize}()},
\code{\link{meta.character}()},
\code{\link{meta.numeric}()},
\code{\link{metaplot.character}()},
\code{\link{metaplot.numeric}()},
\code{\link{metaplot_character}()},
\code{\link{metasuperset}()},
\code{\link{meta}()},
\code{\link{ninput.character}()},
\code{\link{ninput.numeric}()},
\code{\link{ninput}()},
\code{\link{superset.character}()},
\code{\link{superset.numeric}()},
\code{\link{superset}()},
\code{\link{superspec.character}()},
\code{\link{superspec.numeric}()},
\code{\link{superspec}()}
}
\concept{superset}
