% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likebut.R
\name{specfile}
\alias{specfile}
\title{Identify the Data Specification File for a Model}
\usage{
specfile(x, ..., opt = getOption("project"), project = if (is.null(opt))
  getwd() else opt)
}
\arguments{
\item{x}{the model name}

\item{...}{passed arguments}

\item{opt}{alternative specification of project directory}

\item{project}{direct specification of project directory}
}
\value{
character
}
\description{
Identifies the data specification file associated with the datafile used by a model.
}
