% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{specfile}
\alias{specfile}
\title{Identify Specfile}
\usage{
specfile(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{dots}
}
\description{
Identifies specfile.
}
\seealso{
Other path: 
\code{\link{datafile.character}()},
\code{\link{datafile.numeric}()},
\code{\link{datafile}()},
\code{\link{modeldir}()},
\code{\link{modelfile}()},
\code{\link{modelpath.character}()},
\code{\link{modelpath.numeric}()},
\code{\link{modelpath}()},
\code{\link{psn_options}()},
\code{\link{specfile.character}()},
\code{\link{specfile.numeric}()}
}
\concept{path}
\keyword{internal}
