% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{[[.model}
\alias{[[.model}
\title{Select model Element}
\usage{
\method{[[}{model}(x, ..., drop = TRUE)
}
\arguments{
\item{x}{model}

\item{...}{dots}

\item{drop}{passed to element select}
}
\value{
element
}
\description{
Selects model element.
}
\seealso{
Other as.model: 
\code{\link{[.model}()},
\code{\link{as.model.numeric}()},
\code{\link{as.model}()},
\code{\link{read.model}()},
\code{\link{write.model}()}
}
\concept{as.model}
\keyword{internal}
