% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superset.R
\name{superset.numeric}
\alias{superset.numeric}
\title{Coerce to Superset from Numeric}
\usage{
\method{superset}{numeric}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Coerces to superset from numeric.
}
\details{
Converts number to character and revisits generic.
}
\seealso{
\code{\link{superset.character}}
}
\keyword{internal}
