% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superset.R
\name{superspec}
\alias{superspec}
\title{Create Specification for Model Inputs and Outputs}
\usage{
superspec(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Create a specification for the result of superset().
}
\seealso{
\code{\link{superspec.character}}

Other superset: 
\code{\link{generalize}()},
\code{\link{meta.character}()},
\code{\link{meta.numeric}()},
\code{\link{metaplot.character}()},
\code{\link{metaplot.numeric}()},
\code{\link{metaplot_character}()},
\code{\link{metasuperset}()},
\code{\link{meta}()},
\code{\link{ninput.character}()},
\code{\link{ninput.numeric}()},
\code{\link{ninput}()},
\code{\link{shuffle}()},
\code{\link{superset.character}()},
\code{\link{superset.numeric}()},
\code{\link{superset}()},
\code{\link{superspec.character}()},
\code{\link{superspec.numeric}()}
}
\concept{superset}
