% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{tweak.init}
\alias{tweak.init}
\title{Tweak Init}
\usage{
\method{tweak}{init}(x, sd = 0.13, digits = 3, ...)
}
\arguments{
\item{x}{object}

\item{sd}{numeric}

\item{digits}{integer}

\item{...}{dots}
}
\value{
init
}
\description{
Tweaks init.
}
\seealso{
Other tweak: 
\code{\link{tweak.default}()},
\code{\link{tweak.inits}()},
\code{\link{tweak.model}()},
\code{\link{tweak}()}
}
\concept{tweak}
\keyword{internal}
