% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.init.R
\name{tweak}
\alias{tweak}
\title{Tweak something}
\usage{
tweak(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{dots}
}
\description{
Tweaks something.
}
\seealso{
Other tweak: \code{\link{tweak.initList}},
  \code{\link{tweak.init}}, \code{\link{tweak.nmctl}}
}

