% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{updated.numeric}
\alias{updated.numeric}
\title{Create the Updated Version of Numeric}
\usage{
\method{updated}{numeric}(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{dots}
}
\description{
Creates the updated version of numeric by coercing to character.
}
\seealso{
Other updated: 
\code{\link{updated.character}()},
\code{\link{updated}()}
}
\concept{updated}
\keyword{internal}
