% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partab.R
\name{val_name}
\alias{val_name}
\title{Create Model Parameter Table from Xml_document}
\usage{
val_name(x, xpath, param, moment, ...)
}
\arguments{
\item{x}{xml_document}

\item{xpath}{character}

\item{param}{character}

\item{moment}{character}

\item{...}{passed arguments}
}
\value{
data.frame
}
\description{
Creates a model parameter table from xml_document.
}
\details{
The function evaluates the xpath expression, 
supplying arbitrary names for parameter and moment of parameter.
}
\concept{val_name}
\keyword{internal}
