% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpath.R
\name{xpath.default}
\alias{xpath.default}
\title{Evaluate xpath Expression in Default Context}
\usage{
\method{xpath}{default}(x, ...)
}
\arguments{
\item{x}{default}

\item{...}{passed arguments}
}
\value{
vector
}
\description{
Coerces x to xml_document and evaluates.
}
\seealso{
Other xpath: 
\code{\link{as.xml_document}()},
\code{\link{as.xml_document.character}()},
\code{\link{as.xml_document.numeric}()},
\code{\link{xpath}()},
\code{\link{xpath.xml_document}()}
}
\concept{xpath}
