% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpath.R
\name{xpath}
\alias{xpath}
\title{Evaluate Xpath Expression}
\usage{
xpath(x, ...)
}
\arguments{
\item{x}{xml_document}

\item{...}{passed arguments}
}
\description{
Evaluates an xpath expression.
}
\details{
The resulting nodeset is scavenged for text, and coerced to best of numeric or character.
}
\examples{
library(magrittr)
options(project = system.file('project/model',package='nonmemica'))
1001 \%>\% xpath('//etashrink/row/col')
}
\seealso{
Other xpath: 
\code{\link{as.xml_document}()},
\code{\link{as.xml_document.character}()},
\code{\link{as.xml_document.numeric}()},
\code{\link{xpath.default}()},
\code{\link{xpath.xml_document}()}
}
\concept{xpath}
